Refactor modules: Remove hardcoded magical array size. [4/4]
ScienceData
SportData
SystemData
VideoGameData
VehicleData
WeatherData
WordData
