+++ src/modules/videoGame/VideoGameData.h
@@ -7,9 +7,3591 @@ namespace faker
 {
 namespace videoGame
 {
-extern const std::array<std::string_view, 3152> videoGameNames;
-extern const std::array<std::string_view, 28> videoGameGenres;
-extern const std::array<std::string_view, 9> platforms;
-extern const std::array<std::string_view, 392> studioNames;
+const auto videoGameNames = std::to_array<std::string_view>({
+    "#killallzombies",
+    ".hack//G.U. Last Recode",
+    "007: Agent Under Fire",
+    "007: Everything or Nothing",
+    "007: From Russia with Love",
+    "007: Nightfire",
+    "10 Second Ninja X",
+    "1001 Spikes",
+    "100ft Robot Golf",
+    "101 Ways to Die",
+    "11-11: Memories Retold",
+    "13 Sentinels: Aegis Rim",
+    "140",
+    "1552: Tenka Tairan",
+    "187 Ride or Die",
+    "1941: Counter Attack",
+    "1943 Kai",
+    "1979 Revolution: Black Friday",
+    "198X",
+    "2002 FIFA World Cup",
+    "2064: Read Only Memories",
+    "20XX",
+    "25 To Life",
+    "2Dark",
+    "3 Minutes to Midnight",
+    "39 Days to Mars",
+    "3D Billiards",
+    "3D MiniGolf",
+    "428: Shibuya Scramble",
+    "4x4 EVO 2",
+    "5 Star Wrestling: ReGenesis",
+    "50 Cent: Bulletproof",
+    "60 Parsecs!",
+    "60 Seconds!",
+    "7 Days to Die",
+    "7th Sector",
+    "8 to Glory",
+    "8-Bit Armies",
+    "8-Bit Hordes",
+    "8-Bit Invaders",
+    "8-bit Adventure Anthology: Volume I",
+    "88 Heroes",
+    "9 Monkeys of Shaolin",
+    "911 Operator",
+    "99Vidas",
+    "9th Dawn III",
+    "A Bastard's Tale",
+    "A Boy and His Blob",
+    "A Certain Magical Virtual-On",
+    "A Fold Apart",
+    "A Hat in Time",
+    "A Healer Only Lives Twice",
+    "A Hole New World",
+    "A Juggler's Tale",
+    "A Knight's Quest",
+    "A Pixel Story",
+    "A Plague Tale: Innocence",
+    "A Space for the Unbound",
+    "A Tale of Paper",
+    "A Way Out",
+    "A-Train Express",
+    "AER: Memories of Old",
+    "AFL Evolution",
+    "AFL Evolution 2",
+    "AFL Live 2003",
+    "AFL Live 2004",
+    "AFL Live Premiership Edition",
+    "AFL Premiership 2005",
+    "AI-Limit",
+    "AI: The Somnium Files",
+    "AIPD",
+    "AMF Bowling 2004",
+    "AMF Xtreme Bowling 2006",
+    "AND 1 Streetball",
+    "AO Tennis",
+    "AO Tennis 2",
+    "APEX - Racing Evoluzione (PAL)",
+    "ATV Drift and Tricks",
+    "ATV Quad Power Racing 2",
+    "AVICII Invector",
+    "AWAY: The Survival Series",
+    "Aaero",
+    "Aaru's Awakening",
+    "Absolute Drift: Zen Edition",
+    "Absolver",
+    "Abyss Odyssey",
+    "Abyss: The Wraiths of Eden",
+    "Abzû",
+    "Accel World vs. Sword Art Online: Millennium Twilight",
+    "Ace Combat 7: Skies Unknown",
+    "Ace of Seafood",
+    "Aces of the Luftwaffe",
+    "Achtung! Cthulhu Tactics",
+    "Act It Out! A Game of Charades",
+    "Action Henk",
+    "Active Neurons",
+    "Addams Family, The",
+    "Adrift",
+    "Advanced V.G.",
+    "Advent Rising",
+    "Adventure Time: Finn & Jake Investigations",
+    "Adventures of Pip",
+    "Adventures of Scarlet Curiosity",
+    "Aegis Defenders",
+    "Aegis of Earth: Protonovus Assault",
+    "Aeon Must Die!",
+    "Aero Blasters",
+    "AeternoBlade",
+    "AeternoBlade II",
+    "After Burner II",
+    "Afterparty",
+    "Agatha Christie: The ABC Murders",
+    "Age of Wonders: Planetfall",
+    "Agent A: A Puzzle in Disguise",
+    "Agents of Mayhem",
+    "Aggelos",
+    "Aggressive Inline",
+    "Agony",
+    "Aikano: Yukizora no Triangle",
+    "Air Conflicts: Pacific Carriers",
+    "Air Conflicts: Vietnam",
+    "Aircraft Evolution",
+    "Airforce Delta Storm - Deadly Skies (PAL) - Airforce Delta II (JP)",
+    "Airport Simulator 2019",
+    "Airship Q",
+    "Akiba's Beat",
+    "Akiba's Trip: Hellbound & Debriefed",
+    "Akiba's Trip: Undead & Undressed",
+    "Akita Oga Mystery Guide: The Frozen Silverbell Flower",
+    "Akumajō Dracula X: Chi no Rondo",
+    "Alaloth: Champions of The Four Kingdoms",
+    "Alba: A Wildlife Adventure",
+    "Aldynes",
+    "Alekhine's Gun",
+    "Aleste Collection",
+    "Alex Kidd in Miracle World DX",
+    "Alfa Romeo Racing Italiano - SCAR : Squadra Corse Alfa Romeo (PAL)",
+    "Alias",
+    "Alien Crush",
+    "Alien Hominid",
+    "Alien: Isolation",
+    "Alienation",
+    "Aliens Versus Predator: Extinction",
+    "All-Star Baseball 2003",
+    "All-Star Baseball 2004",
+    "All-Star Baseball 2005",
+    "All-Star Fruit Racing",
+    "Alone with You",
+    "Alshark",
+    "Alter Echo",
+    "Alwa's Awakening",
+    "Alwa's Legacy",
+    "Always Sometimes Monsters",
+    "Alzadick",
+    "Amazing Discoveries in Outer Space",
+    "America's Army: Rise of a Soldier",
+    "American Chopper",
+    "American Chopper 2: Full Throttle",
+    "American Fugitive",
+    "American McGee Presents: Scrapland",
+    "Amnesia Collection",
+    "Amnesia: Rebirth",
+    "Amoeba Battle: Microscopic RTS Action",
+    "Among the Sleep",
+    "Amped 2 - Tenku 2 (JP)",
+    "Amped: Freestyle Snowboarding",
+    "Amplitude",
+    "Anarcute",
+    "Ancestors Legacy",
+    "Ancestors: The Humankind Odyssey",
+    "Andre Panza Kick Boxing",
+    "Ane-san",
+    "Anew: The Distant Light",
+    "Angelic Concert (JP)",
+    "AngerForce: Reloaded",
+    "Angry Birds Star Wars",
+    "Anima: Gate of Memories",
+    "Animaniacs: The Great Edgar Hunt",
+    "Anime Studio Story",
+    "Anno: Mutationem",
+    "Anodyne",
+    "Anodyne 2: Return to Dust",
+    "Anoko wa Ore kara Hanarenai",
+    "Anomaly 2",
+    "Anonymous;Code",
+    "Another World: 20th Anniversary Edition",
+    "Anthem",
+    "Antiquia Lost",
+    "Antz Extreme Racing",
+    "Aoi Blink",
+    "Aoi Namida (JP)",
+    "Aokana: Four Rhythm Across the Blue",
+    "Apocalypse",
+    "Apotheon",
+    "Appare! Gateball",
+    "Aqua Moto Racing Utopia",
+    "Aquaman: Battle for Atlantis",
+    "Ar Nosurge DX",
+    "Ara Fell: Enhanced Edition",
+    "Aragami",
+    "Aragami 2",
+    "Arc of Alchemist",
+    "Arcade Classics Anniversary Collection",
+    "Arcade Spirits",
+    "Arcania: The Complete Tale",
+    "Archaica: The Path of Light",
+    "Arctic Thunder",
+    "Area 51",
+    "Arena Football",
+    "Arise: A Simple Story",
+    "Ark: Survival Evolved",
+    "Armed and Dangerous",
+    "Armello",
+    "Army Men: Major Malfunction",
+    "Army Men: Sarge's War",
+    "Arslan: The Warriors of Legend",
+    "Art of Balance",
+    "Arx Fatalis",
+    "Ary and the Secret of Seasons",
+    "Asdivine Dios",
+    "Asdivine Hearts",
+    "Asdivine Hearts II",
+    "Asdivine Kamura",
+    "Asdivine Menace",
+    "Asemblance",
+    "Asemblance: Oversight",
+    "Ashen",
+    "Ashes Cricket",
+    "Assassin's Creed Chronicles: China",
+    "Assassin's Creed Chronicles: India",
+    "Assassin's Creed Chronicles: Russia",
+    "Assassin's Creed III Remastered",
+    "Assassin's Creed IV: Black Flag",
+    "Assassin's Creed Odyssey",
+    "Assassin's Creed Rogue Remastered",
+    "Assassin's Creed Syndicate",
+    "Assassin's Creed Unity",
+    "Assassin's Creed Valhalla",
+    "Assassin's Creed: Origins",
+    "Assassin's Creed: The Ezio Collection",
+    "Assault Android Cactus",
+    "Assault Suit Leynos",
+    "Assetto Corsa",
+    "Assetto Corsa Competizione",
+    "Astebreed",
+    "Asterix & Obelix XXL 2: Mission: Las Vegum",
+    "Asterix & Obelix XXL 3: The Crystal Menhir",
+    "Asterix & Obelix XXL: Romastered",
+    "Asterix & Obelix: Slap Them All!",
+    "Astralius",
+    "Astroneer",
+    "Asuka 120% Maxima",
+    "Atari Anthology",
+    "Atari Flashback Classics: Volume 1",
+    "Atari Flashback Classics: Volume 2",
+    "Atari Flashback Classics: Volume 3",
+    "Atelier Ayesha: The Alchemist of Dusk DX",
+    "Atelier Escha & Logy: Alchemists of the Dusk Sky DX",
+    "Atelier Firis: The Alchemist and the Mysterious Journey",
+    "Atelier Lulua: The Scion of Arland",
+    "Atelier Lydie & Suelle: Alchemists of the Mysterious Painting",
+    "Atelier Meruru: The Apprentice of Arland",
+    "Atelier Rorona: The Alchemist of Arland",
+    "Atelier Ryza 2: Lost Legends & the Secret Fairy",
+    "Atelier Ryza: Ever Darkness & the Secret Hideout",
+    "Atelier Shallie: Alchemists of the Dusk Sea DX",
+    "Atelier Sophie: The Alchemist of the Mysterious Book",
+    "Atelier Totori: The Adventurer of Arland",
+    "Atomic Heart",
+    "Atomic Robo-Kid Special",
+    "Atomicrops",
+    "Attack on Titan",
+    "Attack on Titan 2",
+    "Attractio",
+    "Aurora Quest: Otaku no Seiza in Another World",
+    "Auto Chess",
+    "Auto Crusher Palladium",
+    "Auto Modellista",
+    "Autumn's Journey",
+    "Avatar: The Last Airbender",
+    "Aven Colony",
+    "Avenger",
+    "Away: Journey to the Unexpected",
+    "Awesomenauts Assemble",
+    "Axiom Verge",
+    "Azkend 2: The World Beneath",
+    "Aztech: Forgotten Gods",
+    "Azur Lane: Crosswave",
+    "Azure Reflections",
+    "Azure Striker Gunvolt: Striker Pack",
+    "Azurik: Rise of Perathia",
+    "BMX XXX",
+    "BPM: Bullets Per Minute",
+    "BQM: BlockQuest Maker",
+    "Babel",
+    "Baby Jo the Superhero",
+    "Babylon's Fall",
+    "Back to Bed",
+    "Back to the Future: The Game – 30th Anniversary Edition",
+    "Backbone",
+    "Backgammon Blitz",
+    "Backyard Wrestling 2: There Goes the Neighborhood",
+    "Backyard Wrestling: Don't Try This at Home",
+    "Bad Boys: Miami Takedown - Bad Boys II (PAL)",
+    "Bad North",
+    "Badland: Game of the Year Edition",
+    "Baja: Edge of Control HD",
+    "Bake 'n Switch",
+    "Bakuden: Unbalanced Zone",
+    "Balan Wonderworld",
+    "Baldur's Gate II: Enhanced Edition",
+    "Baldur's Gate: Dark Alliance",
+    "Baldur's Gate: Dark Alliance II",
+    "Baldur's Gate: Enhanced Edition",
+    "Baldur's Gate: Siege of Dragonspear",
+    "Ballistix",
+    "Banner of the Maid",
+    "Barbarian",
+    "Barbie Horse Adventures: Wild Horse Rescue",
+    "Bari Bari Densetsu",
+    "Barunba",
+    "Basement Crawl",
+    "Bass Pro Shops Trophy Hunter 2007",
+    "Bass Pro Shops: Trophy Bass 2007",
+    "Basted",
+    "Bastion",
+    "Batman",
+    "Batman Begins",
+    "Batman Vengeance",
+    "Batman: Arkham Knight",
+    "Batman: Dark Tomorrow",
+    "Batman: Return to Arkham",
+    "Batman: Rise of Sin Tzu",
+    "Batman: The Enemy Within",
+    "Batman: The Telltale Series",
+    "Battalion 1944",
+    "Battle Ace",
+    "Battle Chasers: Nightwar",
+    "Battle Engine Aquila",
+    "Battle Field '94 in Tokyo Dome",
+    "Battle Garegga Rev.2016",
+    "Battle Lode Runner",
+    "Battle Princess Madelyn",
+    "Battle Royale",
+    "Battle Worlds: Kronos",
+    "Battle of the Bulge",
+    "Battlefield 1",
+    "Battlefield 2: Modern Combat",
+    "Battlefield 4",
+    "Battlefield Hardline",
+    "Battlefield V",
+    "Battleship",
+    "Battlestar Galactica",
+    "Battlestar Galactica: Deadlock",
+    "Battlezone",
+    "Batu Ta Batu",
+    "Bayonetta",
+    "Bazar dé Gozarre no Game de Gozāru",
+    "Bear With Me: The Lost Robots",
+    "Bears Can't Drift!?",
+    "Beast Quest",
+    "Beat Down: Fists of Vengeance",
+    "Beautiful Desolation",
+    "Bedlam",
+    "Bee Simulator",
+    "Beholder 2",
+    "Beholder: Complete Edition",
+    "Below",
+    "Ben 10",
+    "Ben 10: Power Trip",
+    "Bendy and the Ink Machine",
+    "Benkei Gaiden",
+    "Berserk and the Band of the Hawk",
+    "Beyond Blue",
+    "Beyond Eyes",
+    "Beyond Good & Evil",
+    "Beyond Shadowgate",
+    "Beyond a Steel Sky",
+    "Beyond: Two Souls",
+    "Bicycle Casino",
+    "Big Bash Boom",
+    "Big Bumpin'",
+    "Big Mutha Truckers",
+    "Big Mutha Truckers 2",
+    "Big Pharma",
+    "Bikkuriman Daijikai",
+    "Bikkuriman World",
+    "Binaries",
+    "BioShock: The Collection",
+    "Biomutant",
+    "Bionicle",
+    "Biped",
+    "Birthday of Midnight",
+    "Birthdays the Beginning",
+    "Bishōjo Senshi Sailor Moon",
+    "Bishōjo Senshi Sailor Moon Collection",
+    "Bistro Cupid (JP)",
+    "Bistro Cupid 2",
+    "Bit.Trip",
+    "Black",
+    "Black & White Bushido",
+    "Black Desert",
+    "Black Hole Assault",
+    "Black Legend",
+    "Black Paradox",
+    "Black Stone: Magic & Steel - Ex-Chaser (JP)",
+    "Black the Fall",
+    "Blackguards 2",
+    "Blackhole: Complete Edition",
+    "Blacksad: Under the Skin",
+    "Blacksea Odyssey",
+    "Blackwood Crossing",
+    "Blade Arcus from Shining EX",
+    "Blade Assault",
+    "Blade Ballet",
+    "Blade II",
+    "Blade Runner: Enhanced Edition",
+    "Blade Strangers",
+    "Bladed Fury",
+    "Bladestorm: Nightmare",
+    "Blair Witch",
+    "Blasphemous",
+    "Blast 'Em Bunnies",
+    "Blast Zone! Tournament",
+    "Blaster Master Zero",
+    "Blaster Master Zero 2",
+    "Blaster Master Zero 3",
+    "Blaster Master Zero Trilogy: MetaFight Chronicle",
+    "BlazBlue: Central Fiction",
+    "BlazBlue: Chrono Phantasma Extend",
+    "BlazBlue: Cross Tag Battle",
+    "Blazerush",
+    "Blazing Angels: Squadrons of WWII",
+    "Blazing Beaks",
+    "Blazing Chrome",
+    "Blazing Lazers - •GunhedJP",
+    "Bleed",
+    "Bleed 2",
+    "Blind Fate: Edo no Yami",
+    "Blinx 2: Masters of Time and Space - Blinx 2: Battle of Time and Space (JP)",
+    "Blinx: The Time Sweeper",
+    "Blitz: The League",
+    "Block-a-Pix Deluxe",
+    "Blood Bowl 2",
+    "Blood Bowl 3",
+    "Blood Gear",
+    "Blood Omen 2",
+    "Blood Wake",
+    "BloodRayne",
+    "BloodRayne 2",
+    "Bloodborne",
+    "Bloodroots",
+    "Bloodstained: Curse of the Moon",
+    "Bloodstained: Curse of the Moon 2",
+    "Bloodstained: Ritual of the Night",
+    "Bloody Roar Extreme",
+    "Bloody Wolf - •Narazumono Sentō Butai Bloody WolfJP",
+    "Bloody Zombies",
+    "Bloons TD 5",
+    "BlowOut",
+    "Blue Estate",
+    "Blue Fire",
+    "Blue Reflection",
+    "Blue Rider",
+    "Blues and Bullets",
+    "Boiling Bolt",
+    "Bokosuka Wars II",
+    "Bokuhime Project",
+    "Bomber Crew",
+    "Bomberman",
+    "Bomberman '93",
+    "Bomberman '94",
+    "Bomberman Users Battle",
+    "Bonanza Bros.",
+    "Bonk 3: Bonk's Big Adventure - •PC Genjin 3JP",
+    "Bonk 3: Bonk's Big Adventure CD-ROM² - •PC Genjin 3JP",
+    "Bonk's Adventure - •PC GenjinJP",
+    "Bonk's Revenge - •PC Genjin 2JP",
+    "Book of Demons",
+    "Borderlands 3",
+    "Borderlands: Game of the Year Edition",
+    "Borderlands: The Handsome Collection",
+    "Bouken Danshaku Don Sun=Heart-hen",
+    "Bound",
+    "Bound by Flame",
+    "Boundary",
+    "Boundless",
+    "Bounty Battle",
+    "Bow to Blood: Last Captain Standing",
+    "Boxyboy - •Sōkoban WorldJP",
+    "Braid Anniversary Edition",
+    "Brandish",
+    "Braveknight",
+    "Braveland Trilogy",
+    "Bravoman - •Chōzetsurin-jin BerabōmanJP",
+    "Brawl",
+    "Brawlout",
+    "Break In",
+    "Break Nine: World Billiards Tournament",
+    "Breakdown",
+    "Breakers Collection",
+    "Breeders' Cup World Thoroughbred Championships",
+    "Brian Lara International Cricket 2005",
+    "Bridge Constructor",
+    "Bridge Constructor Portal",
+    "Bridge Constructor Stunts",
+    "Bridge Constructor: The Walking Dead",
+    "Brief Battles",
+    "Brigandine: The Legend of Runersia",
+    "Bright Memory: Infinite",
+    "Broforce",
+    "Broken Age: The Complete Adventure",
+    "Broken Delusion",
+    "Broken Sword 5: The Serpent's Curse",
+    "Broken Sword: The Sleeping Dragon",
+    "Brothers in Arms: Earned in Blood",
+    "Brothers in Arms: Road to Hill 30",
+    "Brothers: A Tale of Two Sons",
+    "Browning",
+    "Bruce Lee: Quest of the Dragon",
+    "Brutal",
+    "Brute Force",
+    "Bubble Bobble 4 Friends: The Baron is Back",
+    "Bubblegum Crash",
+    "Bubsy: Paws on Fire!",
+    "Bubsy: The Woolies Strike Back",
+    "Bucket Knight",
+    "Bud Spencer & Terence Hill: Slaps and Beans",
+    "Buffy the Vampire Slayer",
+    "Buffy the Vampire Slayer: Chaos Bleeds",
+    "Bug Fables: The Everlasting Sapling",
+    "BugsBox",
+    "Bugsnax",
+    "Builder Land",
+    "Buildings Have Feelings Too!",
+    "Bulb Boy",
+    "Bulletstorm: Full Clip Edition",
+    "Bullfight Ring no Hasha",
+    "Burai",
+    "Burai II",
+    "Buried Stars",
+    "Burning Angels",
+    "Burnout",
+    "Burnout 2: Point of Impact",
+    "Burnout 3: Takedown",
+    "Burnout Paradise Remastered",
+    "Burnout Revenge",
+    "Bus Driver Simulator",
+    "Bus Simulator 18",
+    "Bus Simulator 21",
+    "Bush Hockey League",
+    "Buster Bros. - •Pomping WorldJP",
+    "Butcher",
+    "C.A.T.: Cyber Attack Team",
+    "CAL II",
+    "CD Battle: Hikari no Yūshatachi",
+    "CSI: Crime Scene Investigation",
+    "Cabela's Big Game Hunter 2005 Adventures",
+    "Cabela's Dangerous Hunts",
+    "Cabela's Dangerous Hunts 2",
+    "Cabela's Deer Hunt: 2004 Season",
+    "Cabela's Deer Hunt: 2005 Season",
+    "Cabela's Outdoor Adventures",
+    "Cadash",
+    "Cafeteria Nipponica",
+    "Cake Bash",
+    "Cal III",
+    "Caladrius Blaze",
+    "Call of Cthulhu: Dark Corners of the Earth",
+    "Call of Cthulhu: The Official Video Game",
+    "Call of Duty 2: Big Red One",
+    "Call of Duty 3",
+    "Call of Duty: Advanced Warfare",
+    "Call of Duty: Black Ops Cold War",
+    "Call of Duty: Black Ops III",
+    "Call of Duty: Black Ops IIII",
+    "Call of Duty: Finest Hour",
+    "Call of Duty: Ghosts",
+    "Call of Duty: Infinite Warfare",
+    "Call of Duty: Modern Warfare",
+    "Call of Duty: Modern Warfare 2 Campaign Remastered",
+    "Call of Duty: Modern Warfare Remastered",
+    "Call of Duty: WWII",
+    "Call of Duty: Warzone",
+    "Calvino Noir",
+    "Camp California",
+    "Can't Drive This",
+    "Candle: The Power of the Flame",
+    "Candlelight",
+    "Candleman: The Complete Journey",
+    "Cannon Brawl",
+    "Capcom Arcade Stadium",
+    "Capcom Beat 'Em Up Bundle",
+    "Capcom Classics Collection Vol. 1",
+    "Capcom Classics Collection Vol. 2",
+    "Capcom Fighting Evolution - Capcom Fighting Jam (PAL) (JP)",
+    "Capcom vs. SNK 2 EO",
+    "Capsule Force",
+    "Captain Tsubasa: Rise of New Champions",
+    "Car Mechanic Simulator 2018",
+    "CarX Drift Racing Online",
+    "Card Angels",
+    "Cardpocalypse",
+    "Carmageddon: Max Damage",
+    "Carmen Sandiego: The Secret of the Stolen Drums",
+    "Cars",
+    "Cars 3: Driven to Win",
+    "Carto",
+    "Cartoon Network: Battle Crashers",
+    "Carve",
+    "Casey Powell Lacrosse 16",
+    "Castle Crashers Remastered",
+    "CastleStorm II",
+    "CastleStorm: Definitive Edition",
+    "Castles",
+    "Castlevania Anniversary Collection",
+    "Castlevania Requiem",
+    "Castlevania: Curse of Darkness - Akumajo Dracula: Yami no Juin (JP)",
+    "Cat Quest",
+    "Cat Quest II",
+    "Catherine: Full Body",
+    "Catlateral Damage",
+    "Catwoman",
+    "Cel Damage",
+    "Cel Damage HD",
+    "Celebrity Deathmatch",
+    "Celeste",
+    "Chained Echoes",
+    "Chambara",
+    "Champion Wrestler",
+    "Champions Forever",
+    "Championship Bowling",
+    "Championship Manager 2006",
+    "Championship Manager 5",
+    "Championship Manager: Season 01/02",
+    "Championship Manager: Season 02/03",
+    "Championship Rally",
+    "Chaos Code: New Sign of Catastrophe",
+    "Chaos on Deponia",
+    "Chaos;Child",
+    "Chariot",
+    "Charlie and the Chocolate Factory",
+    "Chase H.Q. - •Taito Chase HQJP",
+    "Chase: Hollywood Stunt Driver",
+    "Chasm",
+    "Checkers",
+    "Chernobylite",
+    "Chess Ultra",
+    "Chessmaster 10th Edition",
+    "Chew Man Fu - •Be BallJP",
+    "Chibi Maruko-chan: Quiz de Piihira",
+    "Chicago Enforcer",
+    "Chicken Little",
+    "Chicken Police: Paint it Red!",
+    "Chiki Chiki Boys",
+    "Chikuden-ya Toubei",
+    "Child of Light",
+    "Children of Morta",
+    "Children of Zodiarcs",
+    "Chime Sharp",
+    "Chimparty",
+    "China Warrior - •The Kung FuJP",
+    "Chivalry 2",
+    "Chivalry: Medieval Warfare",
+    "Cho Aniki",
+    "Chocobo's Mystery Dungeon Every Buddy!",
+    "Chop is Dish",
+    "Chorus: Rise As One",
+    "Chroma Squad",
+    "ChromaGun",
+    "Chronicles of Teddy: Harmony of Exidus",
+    "Chronos: Before the Ashes",
+    "Chō Jikū Yōsai Macross 2036",
+    "Chō Jikū Yōsai Macross: Eien no Love Song",
+    "Ciel Nosurge DX",
+    "Cinders",
+    "Circuit Superstars",
+    "Circuits",
+    "Circus Lido",
+    "Circus Maximus: Chariot Wars",
+    "Citadel: Forged with Fire",
+    "Cities: Skylines",
+    "Citizens of Earth",
+    "Citizens of Space",
+    "City Hunter",
+    "City Shrouded in Shadow",
+    "Civilization VI",
+    "Cladun Returns: This is Sengoku!",
+    "Claire: Extended Cut",
+    "Clannad",
+    "Classified: The Sentinel Crisis",
+    "Claybook",
+    "Clockwork Aquario",
+    "Clockwork Tales: Of Glass and Ink",
+    "Close Combat: First to Fight",
+    "Close to the Sun",
+    "Cloudpunk",
+    "Club Football",
+    "Club Football 2005",
+    "Clustertruck",
+    "Cobra II: Densetsu no Otoko",
+    "Cobra Kai: The Karate Kid Saga Continues",
+    "Cobra: Kokuryū-Ō no Densetsu",
+    "Code Vein",
+    "Codename: Kids Next Door - Operation: V.I.D.E.O.G.A.M.E.",
+    "Coffee Talk",
+    "Coffin Dodgers",
+    "Cogen: Sword of Rewind",
+    "Cold Fear",
+    "Cold War",
+    "Colin McRae Rally 04",
+    "Colin McRae Rally 2005",
+    "Colin McRae Rally 3",
+    "College Hoops 2K6",
+    "College Hoops 2K7",
+    "Color Wars",
+    "Colt Canyon",
+    "Columns",
+    "Combat Elite: WWII Paratroopers",
+    "Combat: Task Force 121",
+    "Comet Crash 2",
+    "Commander Cherry's Puzzled Journey",
+    "Commandos 2 - HD Remaster",
+    "Commandos 2: Men of Courage",
+    "Commandos: Strike Force",
+    "Conan",
+    "Conan Chop Chop",
+    "Conan Exiles",
+    "Conarium",
+    "Concept Destruction",
+    "Conception Plus: Maidens of the Twelve Stars",
+    "Concrete Genie",
+    "Conflict: Desert Storm",
+    "Conflict: Desert Storm II: Back to Baghdad",
+    "Conflict: Global Terror",
+    "Conflict: Vietnam",
+    "Conga Master",
+    "Conker: Live & Reloaded",
+    "Conspiracy: Weapons of Mass Destruction",
+    "Constantine",
+    "Construction Simulator 2: Console Edition",
+    "Construction Simulator 3: Console Edition",
+    "Constructor HD",
+    "Constructor Plus",
+    "Contra Anniversary Collection",
+    "Contra: Rogue Corps",
+    "Contrast",
+    "Control",
+    "Convallaria",
+    "Convoy: A Tactical Roguelike",
+    "Cook, Serve, Delicious! 2",
+    "Cook, Serve, Delicious! 3",
+    "Corpse Party: Blood Covered Repeated Fear",
+    "Corvette",
+    "Coryoon",
+    "Cosmic Fantasy 2 - •Cosmic Fantasy 2: Bōken Shōnen VanJP",
+    "Cosmic Fantasy 3: Bōken Shōnen Rei",
+    "Cosmic Fantasy 4: Ginga Shōnen Densetsu: Gekitouhen",
+    "Cosmic Fantasy 4: Ginga Shōnen Densetsu: Totsunyūhen",
+    "Cosmic Fantasy: Bōken Shōnen Yū",
+    "Cosmic Star Heroine",
+    "Costume Quest 2",
+    "Cotton Reboot!",
+    "Cotton: Fantastic Night Dreams - •CottonJP",
+    "Counter-Strike",
+    "CounterSpy",
+    "Cozy Groove",
+    "Crash 'n' Burn",
+    "Crash Bandicoot 4: It's About Time",
+    "Crash Bandicoot N. Sane Trilogy",
+    "Crash Bandicoot: The Wrath of Cortex",
+    "Crash Nitro Kart",
+    "Crash Tag Team Racing",
+    "Crash Team Racing Nitro-Fueled",
+    "Crash Twinsanity",
+    "Cratermaze - •Doraemon: Meikyū DaisakusenJP",
+    "Crawl",
+    "Crayola Scoot",
+    "Crazy Strike Bowling EX",
+    "Crazy Taxi 3: High Roller",
+    "Creaks",
+    "Creature in the Well",
+    "Cricket 19",
+    "Cricket 2005",
+    "Crime Life: Gang Wars",
+    "Crimson Sea",
+    "Crimson Skies: High Road to Revenge",
+    "Cris Tales",
+    "Cross Wiber: Cyber Combat Police",
+    "CrossCode",
+    "CrossKrush",
+    "Crossing Souls",
+    "Crouching Tiger, Hidden Dragon",
+    "Crown Trick",
+    "Crows: Burning Edge",
+    "Crusty Demons",
+    "Crypt of the NecroDancer",
+    "Cryptark",
+    "Crysis Remastered",
+    "Crystal Crisis",
+    "Crystar",
+    "Cube Life: Island Survival HD",
+    "Cubers: Arena",
+    "Cuphead",
+    "Curious Expedition",
+    "Curious George",
+    "Curse of the Dead Gods",
+    "Curse: The Eye of Isis",
+    "Curses 'N Chaos",
+    "Cyber City Oedo: Kemono no Alignment",
+    "Cyber Core",
+    "Cyber Cross",
+    "Cyber Dodge",
+    "Cyber Knight",
+    "Cyber Troopers Virtual-On Masterpiece 1995–2001",
+    "Cyberdimension Neptunia: 4 Goddesses Online",
+    "Cyberpunk 2077",
+    "DARQ: Complete Edition",
+    "DE.JA",
+    "DJMax Respect",
+    "DRIV3R",
+    "Daemon Vector",
+    "Daggerhood",
+    "Dai Makaimura",
+    "Dai Senryaku VII: Modern Military Tactics - Daisenryaku VII (JP)",
+    "Daichikun Crisis",
+    "Daisenpuu",
+    "Daisenpū Custom",
+    "Daisenryaku II: Campaign Version",
+    "Daisenryaku: Dai Toua Kouboushi 3",
+    "Dakar 18",
+    "Dakar 2: The World's Ultimate Rally",
+    "Damsel",
+    "Dance Dance Revolution Ultramix - Dance Stage Unleashed (PAL)",
+    "Dance Dance Revolution Ultramix 2 - Dance Stage Unleashed 2 (PAL)",
+    "Dance Dance Revolution Ultramix 3 - Dance Stage Unleashed 3 (PAL)",
+    "Dance Dance Revolution Ultramix 4",
+    "Dance: UK",
+    "Dandara",
+    "Danganronpa 1-2 Reload",
+    "Danganronpa Another Episode: Ultra Despair Girls",
+    "Danganronpa Trilogy",
+    "Danganronpa V3: Killing Harmony",
+    "Danger Zone",
+    "Dangerous Golf",
+    "Dangun Feveron",
+    "Darius Alpha",
+    "Darius Cozmic Collection",
+    "Darius Cozmic Revelation",
+    "Darius Plus",
+    "Dariusburst Chronicle Saviours",
+    "Dark Arcana: The Carnival",
+    "Dark Devotion",
+    "Dark Envoy",
+    "Dark Rose Valkyrie",
+    "Dark Souls II",
+    "Dark Souls III",
+    "Dark Souls Remastered",
+    "Dark Summit",
+    "Darkest Dungeon",
+    "Darkestville Castle",
+    "Darksiders Genesis",
+    "Darksiders II: Deathinitive Edition",
+    "Darksiders III",
+    "Darksiders: Warmastered Edition",
+    "Darkwatch",
+    "Darkwing Duck",
+    "Darkwood",
+    "Dave Mirra Freestyle BMX 2",
+    "David Beckham Soccer",
+    "Davis Cup Tennis",
+    "Davis Cup Tennis, The (CD)",
+    "Dawn of the Monsters",
+    "Day of the Tentacle: Remastered Edition",
+    "DayZ",
+    "Daylight",
+    "Daymare: 1998",
+    "Days Gone",
+    "De Blob",
+    "De Blob 2",
+    "Dead Alliance",
+    "Dead Cells",
+    "Dead Island 2",
+    "Dead Island Definitive Edition",
+    "Dead Man's Hand",
+    "Dead Moon",
+    "Dead Nation: Apocalypse Edition",
+    "Dead Rising",
+    "Dead Rising 2",
+    "Dead Rising 2: Off the Record",
+    "Dead Rising 4",
+    "Dead Star",
+    "Dead Synchronicity: Tomorrow Comes Today",
+    "Dead by Daylight",
+    "Dead of the Brain 1 & 2",
+    "Dead or Alive 3",
+    "Dead or Alive 5 Last Round",
+    "Dead or Alive 6",
+    "Dead or Alive Ultimate",
+    "Dead or Alive Xtreme 3",
+    "Dead or Alive Xtreme Beach Volleyball",
+    "Dead or School",
+    "Dead to Rights",
+    "Dead to Rights II",
+    "Deadcore",
+    "Deadlight: Director's Cut",
+    "Deadpool",
+    "Dear Esther",
+    "Death Bringer",
+    "Death Come True",
+    "Death Coming",
+    "Death Crown",
+    "Death Squared",
+    "Death Stranding",
+    "Death end re;Quest",
+    "Death end re;Quest 2",
+    "Death's Gambit",
+    "Deathrow",
+    "Deathsmiles I & II",
+    "Decay of Logos",
+    "Deception IV: The Nightmare Princess",
+    "Deemo: Reborn",
+    "Deep Blue - •Deep Blue: Kaitei ShinwaJP",
+    "Deep Sky Derelicts: Definitive Edition",
+    "Def Jam: Fight for NY",
+    "Defender",
+    "Defenders of Ekron",
+    "Defense Grid 2",
+    "Deformers",
+    "Deiland",
+    "Deko Boko Densetsu",
+    "Deleveled",
+    "Deliver Us the Moon",
+    "Delta Force: Black Hawk Down",
+    "Deltarune",
+    "Demetrios: The Big Cynical Adventure",
+    "Demon Gaze II",
+    "Demon's Tier+",
+    "Demon's Tilt",
+    "Demons Age",
+    "Dengeki Bunko: Fighting Climax Ignition",
+    "Dennou Taisen: DroneZ",
+    "Dennō Tenshi: Digital Angel",
+    "Densha de Go!! Hashirou Yamanote Sen",
+    "Deponia",
+    "Deponia Doomsday",
+    "Descenders",
+    "Desperados III",
+    "Destiny",
+    "Destiny Connect: Tick-Tock Travelers",
+    "Destroy All Humans!",
+    "Destroy All Humans!",
+    "Destroy All Humans! 2",
+    "Detana!! TwinBee",
+    "Detention",
+    "Detroit: Become Human",
+    "Deus Ex: Invisible War",
+    "Deus Ex: Mankind Divided",
+    "Devil Engine",
+    "Devil May Cry 4: Special Edition",
+    "Devil May Cry 5",
+    "Devil May Cry HD Collection",
+    "Devil's Crush - •Devil CrashJP",
+    "Devious Dungeon",
+    "Devious Dungeon 2",
+    "Dex",
+    "Diablo II: Resurrected",
+    "Diablo III",
+    "Diablo IV",
+    "Die Hard",
+    "Die Hard: Vendetta",
+    "Die for Valhalla!",
+    "Digimon Rumble Arena 2",
+    "Digimon Story: Cyber Sleuth",
+    "Digimon Story: Cyber Sleuth – Hacker's Memory",
+    "Digimon Survive",
+    "Digimon World 4 - Dejimon Warudo X (JP)",
+    "Digimon World: Next Order",
+    "Digital Champ: Battle Boxing",
+    "Digital Comic Patlabor: Chapter of Griffon",
+    "Dino Crisis 3",
+    "Dinosaur Hunting",
+    "Dinotopia: The Sunstone Odyssey",
+    "Dirt 4",
+    "Dirt 5",
+    "Dirt Rally",
+    "Dirt Rally 2.0",
+    "Disaster Report 4 Plus: Summer Memories",
+    "Disc Jam",
+    "Disco Elysium: The Final Cut",
+    "Disgaea 1 Complete",
+    "Disgaea 4 Complete+",
+    "Disgaea 5 Complete",
+    "Disgaea 6: Defiance of Destiny",
+    "Dishonored 2",
+    "Dishonored: Death of the Outsider",
+    "Dishonored: Definitive Edition",
+    "Disintegration",
+    "Disney Classic Games: Aladdin and The Lion King",
+    "Disney Infinity 3.0",
+    "Disney Infinity: Marvel Super Heroes",
+    "Disney's Extreme Skate Adventure",
+    "Dissidia Final Fantasy NT",
+    "Distance",
+    "Distraint",
+    "Distraint 2",
+    "Divekick",
+    "Divide",
+    "Divinity: Original Sin",
+    "Divinity: Original Sin II",
+    "DmC: Devil May Cry Definitive Edition",
+    "Do Not Feed the Monkeys",
+    "Doctor Who: The Edge of Reality",
+    "Dogos",
+    "Dokapon UP! Mugen no Roulette",
+    "Doki-Doki Universe",
+    "Don Bradman Cricket 17",
+    "Don Doko Don",
+    "Don't Die, Mr. Robot!",
+    "Don't Starve",
+    "Don't Starve Together",
+    "Donut County",
+    "Doodle Devil",
+    "Doodle God",
+    "Doodle God Evolution",
+    "Doom",
+    "Doom",
+    "Doom 3",
+    "Doom 3",
+    "Doom 3: Resurrection of Evil",
+    "Doom 64",
+    "Doom Eternal",
+    "Doom II",
+    "Door Kickers: Action Squad",
+    "Doraemon Story of Seasons",
+    "Doraemon: Nobita no Dorabian Night",
+    "Doraemon: Nobita no Dorabian Night (CD)",
+    "Double Dragon II: The Revenge",
+    "Double Dungeons",
+    "Double Ring - •W-Ring: The Double RingsJP",
+    "Double-S.T.E.A.L. - The Second Clash",
+    "Doughlings: Arcade",
+    "Download",
+    "Download 2",
+    "Downtown Nekketsu Kōshinkyoku",
+    "Downtown Nekketsu Monogatari",
+    "Downwell",
+    "Dr. Muto",
+    "Dr. Seuss' The Cat in the Hat",
+    "Dragon Age: Inquisition",
+    "Dragon Ball FighterZ",
+    "Dragon Ball Xenoverse",
+    "Dragon Ball Xenoverse 2",
+    "Dragon Ball Z: Idainaru Son Gokou Densetsu",
+    "Dragon Ball Z: Kakarot",
+    "Dragon Ball Z: Sagas",
+    "Dragon EGG!",
+    "Dragon Fantasy: The Black Tome of Ice",
+    "Dragon Fin Soup",
+    "Dragon Half",
+    "Dragon Knight & Graffiti",
+    "Dragon Knight II",
+    "Dragon Knight III",
+    "Dragon Marked For Death",
+    "Dragon Quest Builders",
+    "Dragon Quest Builders 2",
+    "Dragon Quest Heroes",
+    "Dragon Quest Heroes II",
+    "Dragon Quest X",
+    "Dragon Quest XI",
+    "Dragon Quest XI S: Echoes of an Elusive Age - Definitive Edition",
+    "Dragon Saber",
+    "Dragon Slayer: The Legend of Heroes - •Dragon Slayer: Eiyū DensetsuJP",
+    "Dragon Slayer: The Legend of Heroes II",
+    "Dragon Spirit",
+    "Dragon Star Varnir",
+    "Dragon's Crown Pro",
+    "Dragon's Curse - •Adventure IslandJP",
+    "Dragon's Dogma: Dark Arisen",
+    "Dragon's Lair 3D: Return to the Lair",
+    "DragonFangZ: The Rose & Dungeon of Time",
+    "Drake of the 99 Dragons",
+    "Draugen",
+    "Draw a Stickman: Epic 2",
+    "Drawful 2",
+    "Drawn to Death",
+    "Dread Nautical",
+    "Dreamfall Chapters: The Longest Journey",
+    "Dreamfall: The Longest Journey",
+    "Dreams",
+    "Dreamwalker: Never Fall Asleep",
+    "Dreii",
+    "Drihoo (JP)",
+    "Driveclub",
+    "Driveclub Bikes",
+    "Driver: Parallel Lines",
+    "Drop Off - •Drop Rock Hora HoraJP",
+    "Dual Gear",
+    "Duck Game",
+    "Duke Nukem 3D: 20th Anniversary World Tour",
+    "Dungeon Defenders: Awakened",
+    "Dungeon Explorer",
+    "Dungeon Explorer II",
+    "Dungeon Master - •Dungeon Master: Theron's QuestJP",
+    "Dungeon Punks",
+    "Dungeon Village",
+    "Dungeon of the Endless",
+    "DungeonTop",
+    "Dungeons & Dragons: Heroes",
+    "Dungeons & Dragons: Order of the Griffon",
+    "Dungeons 2",
+    "Dungeons 3",
+    "Dungreed",
+    "Dusk Diver",
+    "Dust: An Elysian Tail",
+    "Dustoff Heli Rescue 2",
+    "Dustoff Z",
+    "Dying Light",
+    "Dying Light 2",
+    "Dynastic Hero, The - •Chō Eiyū Densetsu Dynastic HeroJP",
+    "Dynasty Warriors 3 - Shin Sangokumusou 2 (JP)",
+    "Dynasty Warriors 4 - Shin Sangokumusou 3 (JP)",
+    "Dynasty Warriors 5 - Shin Sangokumusou 4 (JP)",
+    "Dynasty Warriors 8: Empires",
+    "Dynasty Warriors 8: Xtreme Legends Complete Edition",
+    "Dynasty Warriors 9",
+    "Dynasty Warriors 9: Empires",
+    "Dynasty Warriors: Godseekers",
+    "Dōkyūsei",
+    "EA Sports UFC",
+    "EA Sports UFC 2",
+    "EA Sports UFC 3",
+    "EA Sports UFC 4",
+    "ELEX",
+    "ESP Ra.De. Psy",
+    "ESPN College Hoops",
+    "ESPN College Hoops 2K5",
+    "ESPN International Winter Sports 2002",
+    "ESPN MLS ExtraTime 2002",
+    "ESPN Major League Baseball",
+    "ESPN NBA 2K5",
+    "ESPN NBA 2Night 2002",
+    "ESPN NBA Basketball",
+    "ESPN NFL 2K5",
+    "ESPN NFL Football",
+    "ESPN NFL PrimeTime 2002",
+    "ESPN NHL 2K5",
+    "ESPN NHL Hockey",
+    "ESPN Winter X-Games Snowboarding 2002",
+    "ESports Life Tycoon",
+    "Earth Atlantis",
+    "Earth Defense Force 4.1",
+    "Earth Defense Force 5",
+    "Earth Defense Force 6",
+    "Earth Defense Force: Iron Rain",
+    "Earth Defense Force: World Brothers",
+    "Earth Wars",
+    "EarthNight",
+    "Earthfall",
+    "Earthlock: Festival of Magic",
+    "Eastshade",
+    "Echo",
+    "Ed, Edd n Eddy: The Mis-Edventures",
+    "Edge of Eternity",
+    "Edna & Harvey: Harvey's New Eyes",
+    "Edna & Harvey: The Breakout",
+    "Efera and Jiliora: The Emblem from Darkness",
+    "Effie",
+    "Egg Mania: Eggstreme Madness",
+    "Eikan wa Kimi ni",
+    "Eitr",
+    "Eiyuden Chronicle: Hundred Heroes",
+    "Eiyū Saigokushi",
+    "Elden Ring",
+    "Electronic Super Joy",
+    "Electronic Super Joy 2",
+    "Elite: Dangerous",
+    "Elliot Quest",
+    "Embers of Mirrim",
+    "Emerald Dragon",
+    "Emily Wants to Play",
+    "Empire of Angels IV",
+    "Empire of Sin",
+    "Enchanted Portals",
+    "Enclave",
+    "Ender Lilies: Quietus of the Knights",
+    "Endless Dungeon",
+    "Endless Fables: Dark Moor",
+    "Energy",
+    "Energy Hook",
+    "England International Football",
+    "Enigmatis 2: The Mists of Ravenwood",
+    "Enigmatis 3: The Shadow of Karkhala",
+    "Enigmatis: The Ghosts of Maple Creek",
+    "Enter the Gungeon",
+    "Enter the Matrix",
+    "Entwined",
+    "Epic Astro Story",
+    "Eragon",
+    "Erica",
+    "Escape Goat 2",
+    "Escape Plan",
+    "Ether One",
+    "Etherborn",
+    "Euro Fishing",
+    "Evan's Remains",
+    "Eve: Valkyrie - Warzone",
+    "Even the Ocean",
+    "Eventide 2: Sorcerer's Mirror",
+    "Eventide 3: Legacy of Legends",
+    "Eventide: Slavic Fable",
+    "Everspace",
+    "Everspace 2",
+    "Everybody's Golf",
+    "Everybody's Gone to the Rapture",
+    "Everything",
+    "Evil Dead: A Fistful of Boomstick",
+    "Evil Dead: Regeneration",
+    "Evil Dead: The Game",
+    "Evil West",
+    "Evolve",
+    "Evotinction",
+    "ExaSkeleton",
+    "Exception",
+    "Exile",
+    "Exile's End",
+    "Exile: Wicked Phenomenon - •Exile IIJP",
+    "Exist Archive",
+    "Exit the Gungeon",
+    "Exophobia",
+    "Extinction",
+    "F-1 CIRCUS'92",
+    "F-1 Dream",
+    "F-1 Pilot",
+    "F.I.S.T.: Forged in Shadow Torch",
+    "F1 2001",
+    "F1 2002",
+    "F1 2015",
+    "F1 2016",
+    "F1 2017",
+    "F1 2018",
+    "F1 2019",
+    "F1 2020",
+    "F1 2021",
+    "F1 Career Challenge",
+    "F1 Circus",
+    "F1 Circus '91",
+    "F1 Circus Special",
+    "F1 Team Simulation Project F",
+    "F1 Triple Battle",
+    "FIA European Truck Racing Championship",
+    "FIFA 06 Soccer",
+    "FIFA 07",
+    "FIFA 14",
+    "FIFA 15",
+    "FIFA 16",
+    "FIFA 17",
+    "FIFA 18",
+    "FIFA 19",
+    "FIFA 20",
+    "FIFA 2005",
+    "FIFA 21",
+    "FIFA Football 2003",
+    "FIFA Football 2004",
+    "FIFA Street",
+    "FIFA Street 2",
+    "FIFA World Cup: Germany 2006",
+    "FILA World Tour Tennis",
+    "Fable",
+    "Fable: The Lost Chapters",
+    "Faceball",
+    "Faeria",
+    "Fairy Fencer F: Advent Dark Force",
+    "Fairy Tail",
+    "Falcon",
+    "Fall Guys: Ultimate Knockout",
+    "Fallen Legion: Revenants",
+    "Fallen Legion: Sins of an Empire",
+    "Fallout 4",
+    "Fallout 76",
+    "Fallout: Brotherhood of Steel",
+    "Family Feud",
+    "Family Guy Video Game!",
+    "Family Mysteries 2: Echoes of Tomorrow",
+    "Family Mysteries: Poisonous Promises",
+    "Fang of Alnam",
+    "Fantastic 4",
+    "Fantasy General II",
+    "Fantasy Zone",
+    "Far Cry 3 Classic",
+    "Far Cry 4",
+    "Far Cry 5",
+    "Far Cry 6",
+    "Far Cry Instincts",
+    "Far Cry Instincts: Evolution",
+    "Far Cry Primal",
+    "Far: Lone Sails",
+    "Farjius no Jakōtei",
+    "Farm Together",
+    "Farming Simulator 15",
+    "Farming Simulator 17",
+    "Farming Simulator 19",
+    "Fast & Furious Crossroads",
+    "Fast Striker",
+    "Fat Princess Adventures",
+    "Fatal Frame - Project Zero (PAL)",
+    "Fatal Frame II: Crimson Butterfly - Project Zero II: Crimson Butterfly (PAL)",
+    "Fate/Extella Link",
+    "Fate/Extella: The Umbral Star",
+    "Fault Milestone One",
+    "Fault Milestone Two Side: Above",
+    "Fault Milestone Two Side: Below",
+    "Faussete Amour",
+    "Fe",
+    "Feist",
+    "Felix the Reaper",
+    "Fell Seal: Arbiter's Mark",
+    "Fenix Furia",
+    "Feudal Alloy",
+    "Fez",
+    "Fibbage",
+    "Fiend Hunter",
+    "Fight Club",
+    "Fight Night 2004",
+    "Fight Night Round 2",
+    "Fight Night: Round 3",
+    "Fight'N Rage",
+    "Fighting EX Layer",
+    "Fighting Run",
+    "Fighting Street",
+    "Figment",
+    "Filthy Lucre",
+    "Final Blaster",
+    "Final Fantasy IX",
+    "Final Fantasy Type-0 HD",
+    "Final Fantasy VII",
+    "Final Fantasy VII Remake",
+    "Final Fantasy VIII Remastered",
+    "Final Fantasy X/X-2 HD Remaster",
+    "Final Fantasy XII: The Zodiac Age",
+    "Final Fantasy XIV: A Realm Reborn",
+    "Final Fantasy XIV: Heavensward",
+    "Final Fantasy XIV: Shadowbringers",
+    "Final Fantasy XIV: Stormblood",
+    "Final Fantasy XV",
+    "Final Fantasy: Crystal Chronicles Remastered",
+    "Final Fight: Streetwise",
+    "Final Horizon",
+    "Final Lap Twin",
+    "Final Match Tennis",
+    "Final Soldier",
+    "Final Zone II",
+    "Finding Nemo",
+    "Fire Blade",
+    "Fire Pro Joshi: Shōmu Chōjo Taisen: Zenjo vs. JWP",
+    "Fire Pro Wrestling 2nd Bout",
+    "Fire Pro Wrestling 3 Legend Bout",
+    "Fire Pro Wrestling Combination Tag",
+    "Fire Pro Wrestling World",
+    "Firewatch",
+    "Fishing Sim World: Pro Tour",
+    "Fishing: Barents Sea",
+    "Fist of the North Star: Lost Paradise",
+    "Five Dates",
+    "Five Nights at Freddy's",
+    "Five Nights at Freddy's 2",
+    "Five Nights at Freddy's 3",
+    "Five Nights at Freddy's 4",
+    "Five Nights at Freddy's: Help Wanted",
+    "Five Nights at Freddy's: Security Breach",
+    "Five Nights at Freddy's: Sister Location",
+    "Flame Over",
+    "Flash Hiders",
+    "Flashback",
+    "Flat Heroes",
+    "FlatOut",
+    "FlatOut 2",
+    "FlatOut 4: Total Insanity",
+    "Flight Academy",
+    "Flinthook",
+    "Flipper Mechanic",
+    "Flipping Death",
+    "Flockers",
+    "Flow",
+    "Flower",
+    "Fluster Cluck",
+    "Football, Tactics & Glory",
+    "For Honor",
+    "For the King",
+    "Forager",
+    "Forced",
+    "Ford Bold Moves Street Racing",
+    "Ford Mustang: The Legend Lives",
+    "Ford Racing 2",
+    "Ford Racing 3",
+    "Ford vs. Chevy",
+    "Foreclosed",
+    "Foregone",
+    "Forgotten Realms: Demon Stone",
+    "Forgotten Worlds",
+    "Forgotton Anne",
+    "Forma.8",
+    "Formation Armed F",
+    "Formation Soccer",
+    "Formation Soccer '95 della Seria A",
+    "Formation Soccer on J-League",
+    "Forza Motorsport",
+    "Foul Play",
+    "Four Sided Fantasy",
+    "Fragments of Him",
+    "Frane: Dragons' Odyssey",
+    "Frantics",
+    "Fray CD",
+    "FreakOut: Calamity TV Show",
+    "Freaky Flyers",
+    "Freddy Spaghetti",
+    "Freedom Fighters",
+    "Freedom Finger",
+    "Freedom Planet",
+    "Freestyle MetalX",
+    "Freestyle Street Soccer",
+    "Friday the 13th: The Game",
+    "Frogger Beyond",
+    "Frogger: Ancient Shadow",
+    "Frostpunk",
+    "Full Metal Panic! Fight! Who Dares Wins",
+    "Full Mojo Rampage",
+    "Full Spectrum Warrior",
+    "Full Spectrum Warrior: Ten Hammers",
+    "Full Throttle Remastered",
+    "Furi",
+    "Furious Karting",
+    "Fury Unleashed",
+    "Fuser",
+    "Fushigi no Gensōkyō 3",
+    "Fushigi no Umi no Nadia",
+    "Fushigi no Yume no Alice",
+    "Futurama",
+    "Future Tactics: The Uprising",
+    "Future Unfolding",
+    "FutureGrind",
+    "Futuridium EP Deluxe",
+    "Fuuraiki 4",
+    "Fuzion Frenzy",
+    "G.I. Joe: Operation Blackout",
+    "Gaia no Monshou",
+    "Gaiflame",
+    "Gain Ground SX",
+    "Gakuen",
+    "Gal Gun 2",
+    "Gal Gun: Double Peace",
+    "Gal Gunvolt",
+    "Gal Gunvolt Burst",
+    "Galacide",
+    "Galaga '90 - •Galaga '88JP",
+    "Galak-Z: The Dimensional",
+    "Galaxy Angel",
+    "Galaxy Deka Gayvan",
+    "Galaxy Fräulein Yuna",
+    "Galaxy Fräulein Yuna 2",
+    "Galaxy of Pen & Paper +1 Edition",
+    "Galleon",
+    "Game Dev Story",
+    "Game Tengoku CruisnMix",
+    "Game of Thrones",
+    "Ganbare! Golf Boys",
+    "Ganbare! Super Strikers",
+    "Gang Beasts",
+    "Garfield Kart: Furious Racing",
+    "Garou: Mark of the Wolves",
+    "Garō Densetsu 2",
+    "Garō Densetsu Special",
+    "Gas Guzzlers Extreme",
+    "Gate of Thunder",
+    "Gauntlet: Dark Legacy",
+    "Gauntlet: Seven Sorrows",
+    "Gauntlet: Slayer Edition",
+    "Gekido: Kintaro's Revenge",
+    "Gekisha Boy",
+    "Gem Smashers",
+    "Gemini: Heroes Reborn",
+    "Gene Troopers",
+    "Generation Zero",
+    "Genji Tsūshin Amedama",
+    "Genma Onimusha",
+    "Genocide",
+    "Genpei Tōma Den",
+    "Gensō No Rondo",
+    "Gensō Rōgoku no Kaleidoscope",
+    "Gensō Tairiku Auleria",
+    "Geometry Wars 3: Dimensions",
+    "Get Even",
+    "Ghost 1.0",
+    "Ghost Blade HD",
+    "Ghost Manor",
+    "Ghost Master: The Gravenville Chronicles",
+    "Ghost Sweeper Mikami",
+    "Ghost of Tsushima",
+    "Ghost of a Tale",
+    "Ghostbusters: The Video Game Remastered",
+    "Ghostrunner",
+    "Ghosts 'n Goblins Resurrection",
+    "Ghoulboy: Dark Sword of Goblin",
+    "Giana Sisters: Twisted Dreams - Director's Cut",
+    "Giga Wrecker Alt.",
+    "Ginga Force",
+    "Ginga Fukei Densetsu Sapphire",
+    "Giraffe and Annika",
+    "Girls und Panzer: Dream Tank Match",
+    "Gladiator: Sword of Vengeance",
+    "Gladius",
+    "Glass Masquerade",
+    "Glass Masquerade 2",
+    "Glitched",
+    "Gnog",
+    "Go! Go! Birdie Chance",
+    "Goat Simulator",
+    "Goblin Commander: Unleash the Horde",
+    "God Eater 2: Rage Burst",
+    "God Eater 3",
+    "God Eater Resurrection",
+    "God Panic: Shijō Saikyō Gundan",
+    "God Wars: Future Past",
+    "God of War",
+    "God of War III Remastered",
+    "God's Trigger",
+    "Gods Will Fall",
+    "Godzilla - •Gojira: Bakutō RetsudenJP",
+    "Godzilla: Destroy All Monsters Melee",
+    "Godzilla: Save the Earth",
+    "Godzilla: The Game",
+    "Going Under",
+    "Gokuraku! Chuuka Daisen",
+    "Golden Axe",
+    "Golden Force",
+    "GoldenEye: Rogue Agent",
+    "Golf Zero",
+    "Golf with Your Friends",
+    "Gomola Speed",
+    "Gone Home",
+    "Gonner",
+    "Gonner 2",
+    "Goodbye Deponia",
+    "Goodbye Volcano High",
+    "Goosebumps: The Game",
+    "Gotcha!",
+    "Gotham Knights",
+    "Gothic Murder: Adventure That Changes Destiny",
+    "Grab the Bottle",
+    "Grabbed by the Ghoulies",
+    "Graceful Explosion Machine",
+    "Gradius",
+    "Gradius II",
+    "Gran Turismo Sport",
+    "Granblue Fantasy Versus",
+    "Granblue Fantasy: Relink",
+    "Grand Ages: Medieval",
+    "Grand Kingdom",
+    "Grand Prix Rock N Racing",
+    "Grand Prix Story",
+    "Grand Theft Auto III",
+    "Grand Theft Auto V",
+    "Grand Theft Auto: San Andreas",
+    "Grand Theft Auto: Vice City",
+    "Gravel",
+    "Graven",
+    "Graveyard Keeper",
+    "Gravity Games Bike: Street. Vert. Dirt.",
+    "Gravity Ghost: Deluxe Edition",
+    "Gravity Heroes",
+    "Gravity Rush 2",
+    "Gravity Rush Remastered",
+    "GreedFall",
+    "Green Hell",
+    "Greg Hastings Tournament Paintball",
+    "Greg Hastings' Tournament Paintball Max'd",
+    "Grey Skies: A War of the Worlds Story",
+    "Grid",
+    "Gridd: Retroenhanced",
+    "Grim Fandango Remastered",
+    "Grim Legends 2: Song of the Dark Swan",
+    "Grim Legends 3: The Dark City",
+    "Grim Legends: The Forsaken Bride",
+    "Grip: Combat Racing",
+    "Gris",
+    "Grood",
+    "Grooverider: Slot Car Thunder",
+    "Group S Challenge - CIRCUS DRIVE (JP)",
+    "Grow Home",
+    "Grow Up",
+    "Guacamelee! 2",
+    "Guacamelee! Super Turbo Championship Edition",
+    "Guard Duty",
+    "Guardians of the Galaxy: The Telltale Series",
+    "Guild of Darksteel",
+    "Guilty Gear",
+    "Guilty Gear Isuka[1]",
+    "Guilty Gear Strive",
+    "Guilty Gear X2#Reload[2]",
+    "Guilty Gear Xrd: Revelator",
+    "Guilty Gear Xrd: Sign",
+    "Guitar Hero Live",
+    "Gulclight TDF-2",
+    "Gulliver Boy",
+    "Gun",
+    "Gun Metal",
+    "GunGriffon: Allied Strike",
+    "Gunboat",
+    "Gundam Breaker 3",
+    "Gundam Versus",
+    "Gungrave G.O.R.E.",
+    "Gunhead",
+    "Gunlord X",
+    "Gunman Clive HD Collection",
+    "Guns of Icarus Alliance",
+    "Guns, Gore and Cannoli",
+    "Guns, Gore and Cannoli 2",
+    "Gunscape",
+    "Gunvalkyrie",
+    "Gunvolt Chronicles: Luminous Avenger iX",
+    "Gyuwanburā Jiko Chūshin Ha: Gekitō Sanjūroku Janshi",
+    "Gyuwanburā Jiko Chūshin Ha: Mahjong Puzzle Collection",
+    "Götzendiener",
+    "H-Hour: World's Elite",
+    "Hakoniwa Company Works",
+    "Half-Life 2",
+    "Halo 2",
+    "Halo 2 Multiplayer Map Pack",
+    "Halo: Combat Evolved",
+    "Hamidashi Creative",
+    "Hammerwatch",
+    "Hanataa ka daka!?",
+    "Hand of Fate",
+    "Hand of Fate 2",
+    "Hard Reset Redux",
+    "Hard West: Ultimate Edition",
+    "Hardcore Mecha",
+    "Hardspace: Shipbreaker",
+    "Hardware: Rivals",
+    "Harry Potter and the Chamber of Secrets",
+    "Harry Potter and the Goblet of Fire",
+    "Harry Potter and the Prisoner of Azkaban",
+    "Harry Potter and the Sorcerer's Stone - Harry Potter and the Philosopher's Stone (PAL)",
+    "Harry Potter: Quidditch World Cup",
+    "Harvest Moon: Light of Hope",
+    "Harvest Moon: Mad Dash",
+    "Harvest Moon: One World",
+    "Has-Been Heroes",
+    "Hataraku Shōjo: Tekipaki Working Love",
+    "Hatoful Boyfriend",
+    "Hatris",
+    "Hatsukoi Monogatari",
+    "Hatsune Miku: Project Diva Future Tone",
+    "Hatsune Miku: Project Diva X",
+    "Haven",
+    "Hawk F-123",
+    "Headhunter Redemption",
+    "Headlander",
+    "Headliner: NoviNews",
+    "Headsnatchers",
+    "Heart&Slash",
+    "Heaven's Vault",
+    "Heavenly Bodies",
+    "Heavy Rain",
+    "Heavy Unit",
+    "Hellblade: Senua's Sacrifice",
+    "Helldivers",
+    "Hellfire S",
+    "Hellfront: Honeymoon",
+    "Hellmut: The Badass from Hell",
+    "Hello Kitty: Roller Rescue - Hello Kitty: Mission Rescue (AS)",
+    "Hellpoint",
+    "Her Majesty's Spiffing",
+    "Here They Lie",
+    "Hero Defense",
+    "Hero Must Die. Again",
+    "Hero Siege",
+    "Heroes of the Pacific",
+    "Heroland",
+    "Hidden Agenda",
+    "Hidden Through Time",
+    "Hide & Dance!",
+    "High Grenadier",
+    "High Heat Major League Baseball 2004",
+    "High Rollers Casino",
+    "Hihō Densetsu: Chris no Bōken",
+    "Himitsu no Hanazono",
+    "Hindsight 20/20",
+    "Hisou Kihei Kai-Serd[b]",
+    "Hit the Ice",
+    "Hitman",
+    "Hitman 2",
+    "Hitman 2: Silent Assassin",
+    "Hitman Go",
+    "Hitman III",
+    "Hitman: Blood Money",
+    "Hitman: Contracts",
+    "HoPiKo",
+    "Hob",
+    "Hogwarts Legacy",
+    "Hohokum",
+    "Hollow Knight",
+    "Holy Potatoes! A Weapon Shop?!",
+    "Holy Potatoes! We're in Space?!",
+    "Holy Potatoes! What the Hell?!",
+    "Home - A Unique Horror Adventure",
+    "Home Free",
+    "Homefront: The Revolution",
+    "Honey in the Sky",
+    "Honey on the Road",
+    "Honō no Dōkyūji: Dodge Danpei",
+    "Hood: Outlaws & Legends",
+    "Horizon Chase Turbo",
+    "Horizon Forbidden West",
+    "Horizon Zero Dawn",
+    "Horned Knight",
+    "Horror Story",
+    "Hot Springs Story",
+    "Hot Wheels Unleashed",
+    "Hot Wheels: Stunt Track Challenge",
+    "Hotel Life: A Resort Simulator",
+    "Hotline Miami",
+    "Hotline Miami 2: Wrong Number",
+    "Hotshot Racing",
+    "House Flipper",
+    "How To Take Off Your Mask Remastered",
+    "How to Survive 2",
+    "How to Survive: Storm Warning Edition",
+    "Hue",
+    "Hulk",
+    "Human Sports Festival",
+    "Human: Fall Flat",
+    "Humanity",
+    "Hummer Badlands",
+    "Hunt: Showdown",
+    "Huntdown",
+    "Hunter: The Reckoning",
+    "Hunter: The Reckoning: Redeemer",
+    "Hyakumonogatari: Honto ni Atta Kowai Hanashi",
+    "Hyper Jam",
+    "Hyper Light Drifter",
+    "Hyper Void",
+    "Hyper Wars",
+    "HyperParasite",
+    "Hypnospace Outlaw",
+    "I Am Bread",
+    "I Am Setsuna",
+    "I, AI",
+    "I-Ninja",
+    "I.Q. Panic",
+    "IHRA Drag Racing 2004",
+    "IHRA Drag Racing: Sportsman Edition",
+    "IHRA Professional Drag Racing 2005",
+    "Ice Age 2: The Meltdown",
+    "Ice Age: Scrat’s Nutty Adventure",
+    "Icewind Dale: Enhanced Edition",
+    "Icey",
+    "Iconoclasts",
+    "Iga Ninden Gaiō",
+    "Ikaruga",
+    "Ikenfell",
+    "Illusion of L'Phalcia",
+    "Image Fight",
+    "Image Fight II",
+    "Immortal Planet",
+    "Immortal Realms: Vampire Wars",
+    "Immortal Redneck",
+    "Immortal: Unchained",
+    "Immortals Fenyx Rising",
+    "Impact Winter",
+    "Impossamole",
+    "In Between",
+    "In Celebration of Violence",
+    "In Nightmare",
+    "In Rays of the Light",
+    "In the Shadows",
+    "Inazuma Eleven: Great Road of Heroes",
+    "Indiana Jones and the Emperor's Tomb",
+    "Indigo Prophecy - Fahrenheit (PAL)",
+    "Indivisible",
+    "Industry Giant II",
+    "IndyCar Series",
+    "IndyCar Series 2005",
+    "Inertial Drift",
+    "Infamous First Light",
+    "Infamous Second Son",
+    "Inferno 2",
+    "Inferno Climber: Reborn",
+    "Infinifactory",
+    "Infinite Minigolf",
+    "Infliction: Extended Cut",
+    "Injustice 2",
+    "Injustice: Gods Among Us Ultimate Edition",
+    "Innocent Tears",
+    "Inoue Mami: Kono Hoshi ni Tatta Hitori no Kimi",
+    "Insane Robots",
+    "Inside",
+    "Inside My Radio",
+    "Inside Pitch 2003",
+    "Insurgency: Sandstorm",
+    "Intellivision Lives!",
+    "International Superstar Soccer 2",
+    "Inuwashi: Urabure Tantei to Ojou-sama Keiji no Ikebukuro Jiken File",
+    "Invector",
+    "Inversus",
+    "Invisible, Inc.",
+    "Invisigun Reloaded",
+    "Ion Fury",
+    "Iris.Fall",
+    "Iron Crypticle",
+    "Iron Harvest",
+    "Iron Phoenix",
+    "Iron Sea Defenders",
+    "Ironcast",
+    "Ironclad Tactics",
+    "Irony Curtain: From Matryoshka with Love",
+    "Is It Wrong to Try to Pick Up Girls in a Dungeon? Infinite Combat",
+    "Island",
+    "It Came From The Desert",
+    "It Takes Two",
+    "It's Quiz Time",
+    "Ittle Dew 2",
+    "J-League Greatest Eleven",
+    "J-Stars Victory Vs+",
+    "J. League Tremendous Soccer '94",
+    "J.B. Harold Murder Club",
+    "J.J. & Jeff - •Kato-chan Ken-chanJP",
+    "Jack Nicklaus' Turbo Golf - •Jack Nicklaus Championship GolfJP",
+    "Jack Nicklaus' Turbo Golf - •Jack Nicklaus' World Golf TourJP",
+    "Jacked",
+    "Jackie Chan's Action Kung Fu - •Jackie ChanJP",
+    "Jade Empire",
+    "James Cameron's Dark Angel",
+    "Jamestown+",
+    "Janshin Densetsu: Quest of Jongmaster",
+    "Jantei Monogatari",
+    "Jantei Monogatari 2: Uchū Tantei Diban: Shutsudō Hen",
+    "Jantei Monogatari 3: Saver Angels",
+    "Japanese Rail Sim: Journey to Kyoto",
+    "Jaseiken Necromancer",
+    "Jaws Unleashed",
+    "Jay and Silent Bob: Mall Brawl",
+    "Jazzpunk: Director's Cut",
+    "Jenny LeClue: Detectivú",
+    "Jeopardy!",
+    "Jet Car Stunts",
+    "Jet Set Knights",
+    "Jet Set Radio Future",
+    "Jetpack Joyride",
+    "Jett: The Far Shore",
+    "Jigoku Meguri",
+    "Jikkyou Powerful Pro Baseball 2016",
+    "Jikkyou World Soccer 2002",
+    "Jim Power",
+    "Jinki Resurrection",
+    "Jinmu Denshō Yaksa",
+    "Jisei: The First Case HD",
+    "JoJo's Bizarre Adventure: Eyes of Heaven",
+    "Jockey's Road",
+    "Joe Dever's Lone Wolf",
+    "John Madden Duo CD-ROM² Football",
+    "John Wick Hex",
+    "Jotun: Valhalla Edition",
+    "Journey",
+    "Journey to the Savage Planet",
+    "Joysound Dive 2",
+    "Judge Dredd: Dredd Vs. Death",
+    "Judgment",
+    "Juiced",
+    "Jumanji: The Video Game",
+    "Jump Force",
+    "Jump King",
+    "Jump Stars",
+    "Jurassic Park: Operation Genesis",
+    "Jurassic World Evolution",
+    "Just Cause",
+    "Just Cause 3",
+    "Just Cause 4",
+    "Just Dance 2021",
+    "Just Deal With It!",
+    "Just Die Already",
+    "Just Sing",
+    "Justice League Heroes",
+    "Jydge",
+    "Jūōki",
+    "Jūōki",
+    "K.O. Seiki Beast Sanjūshi",
+    "Kabuki Ittō Ryōdan",
+    "Kabuki Warriors",
+    "Kagami no Kuni no Legend",
+    "Kaizō Chōjin Shubibinman",
+    "Kaizō Chōjin Shubibinman 3",
+    "Kakuto Chojin",
+    "Kakuto Haō Densetsu Algunos",
+    "Kamen Rider: Battride War Genesis",
+    "Kamen Rider: Memory of Heroez",
+    "Kamiko",
+    "Kandagawa Jet Girls",
+    "Kao the Kangaroo: Round 2",
+    "Karaoke Revolution",
+    "Karaoke Revolution Party",
+    "Karumaruka Circle",
+    "Katamari Damacy Reroll",
+    "Katana Kami: A Way of the Samurai Story",
+    "Kattobi! Takuhai-Kun",
+    "Kawa no Nushizuri: Shizenha",
+    "Kaze Kiri",
+    "Kaze and the Wild Masks",
+    "Keep Talking and Nobody Explodes",
+    "Keith Courage in Alpha Zones - •Mashin Eiyuuden WataruJP",
+    "Kelly Slater's Pro Surfer",
+    "Ken Follett's The Pillars of the Earth",
+    "Kena: Bridge of Spirits",
+    "Kentucky Route Zero: TV Edition",
+    "Kerbal Space Program",
+    "Kerbal Space Program 2",
+    "Kero Blaster",
+    "Ketsui Deathtiny: Kizuna Jigoku Tachi",
+    "Kholat",
+    "KiKi KaiKai",
+    "Kiaidan 00",
+    "Kick Boxing, The",
+    "Kick Off Revival",
+    "KickBeat: Special Edition",
+    "Kickball",
+    "Kikou Heidan J-Phoenix +",
+    "Kill It With Fire",
+    "Kill Switch",
+    "Kill la Kill: If",
+    "Killing Floor 2",
+    "Killing Floor: Double Feature",
+    "Killzone Shadow Fall",
+    "Kin'iro Loveriche",
+    "Kin'iro Loveriche: Golden Time",
+    "King Arthur",
+    "King Oddball",
+    "King Of Casino",
+    "King of Seas",
+    "King's Bounty II",
+    "King's Quest",
+    "Kingdom Come: Deliverance",
+    "Kingdom Hearts III",
+    "Kingdom Hearts: Melody of Memory",
+    "Kingdom Hearts: The Story So Far",
+    "Kingdom Under Fire: Heroes",
+    "Kingdom Under Fire: The Crusaders",
+    "Kingdom of Night",
+    "Kingdom: New Lands",
+    "Kingdom: Two Crowns",
+    "Kingdoms of Amalur: Re-Reckoning",
+    "Kingmaker: Rise to the Throne",
+    "Kingpin: Reloaded",
+    "Kings of Lorn: The Fall of Ebris",
+    "Kisō Louga",
+    "Kitaria Fables",
+    "Kitty Powers' Matchmaker",
+    "Klaus",
+    "Klax",
+    "Knack",
+    "Knack 2",
+    "Knight Rider Special",
+    "Knight's Apprentice: Memorick's Adventures",
+    "Knightin'+",
+    "Knights and Bikes",
+    "Knights of the Temple II",
+    "Knights of the Temple: Infernal Crusade",
+    "Knock-Knock",
+    "Knockout City",
+    "Knockout Kings 2002",
+    "Knot",
+    "Knowledge is Power",
+    "Knowledge is Power: Decades",
+    "Koi",
+    "Koihime Enbu",
+    "Koihime Enbu RyoRaiRai",
+    "Kona",
+    "Kore ga Pro Yakyuu '90",
+    "Kore ga Pro Yakyū '89",
+    "Kotodama: The 7 Mysteries of Fujisawa",
+    "Kowloon's Rhizome: A Day of the Fire",
+    "Kromaia Omega",
+    "Kung Fu Chaos",
+    "Kunio-kun: The World Classics Collection",
+    "Kwaidan: Azuma Manor Story",
+    "Kyūkyoku Tiger",
+    "L-Dis",
+    "L.A. Noire",
+    "L.A. Rush",
+    "LA Cops",
+    "LMA Manager 2003",
+    "LMA Manager 2004",
+    "LMA Manager 2005",
+    "LMA Manager 2006",
+    "La Valeur",
+    "La-Mulana",
+    "La-Mulana 2",
+    "Labyrinth Life",
+    "Labyrinth of Galleria: Coven of Dusk",
+    "Labyrinth of Refrain: Coven of Dusk",
+    "Labyrinth of Zangetsu",
+    "Lady Phantom",
+    "Laid-Back Camp -Virtual- Fumoto Campsite",
+    "Laid-Back Camp -Virtual- Lake Motosu",
+    "Laid-Back Camp: Have a Nice Day!",
+    "Lair of the Clockwork God",
+    "Land of the Dead: Road to Fiddler's Green",
+    "Langrisser I & II HD Remastered",
+    "Langrisser: Hikari no Matsuei",
+    "Lapis x Labyrinth",
+    "Laplace no Ma",
+    "Lara Croft Go",
+    "Lara Croft and the Temple of Osiris",
+    "Largo Winch: Empire Under Threat",
+    "Laser League",
+    "Laserlife",
+    "Last Alert - •Red AlertJP",
+    "Last Armageddon",
+    "Last Day of June",
+    "Last Stop",
+    "Late Shift",
+    "LawBreakers",
+    "Laws of Machine",
+    "Layers of Fear",
+    "Layers of Fear 2",
+    "League of Legends: Wild Rift",
+    "Leap of Fate",
+    "Leaving Lyndow",
+    "Left Alive",
+    "Legacy of Kain: Defiance",
+    "Legend of Hero Tonma",
+    "Legend of Kay Anniversary",
+    "Legend of Mana",
+    "Legend of Xanadu",
+    "Legend of Xanadu II",
+    "Legendary Axe II, The - •Ankoku DensetsuJP",
+    "Legendary Axe, The - •Makyō DensetsuJP",
+    "Legends of Wrestling",
+    "Legends of Wrestling II",
+    "Legion",
+    "Lego Batman 3: Beyond Gotham",
+    "Lego City Undercover",
+    "Lego DC Super-Villains",
+    "Lego Dimensions",
+    "Lego Harry Potter Collection",
+    "Lego Jurassic World",
+    "Lego Marvel Super Heroes",
+    "Lego Marvel Super Heroes 2",
+    "Lego Marvel's Avengers",
+    "Lego Movie Videogame",
+    "Lego Movie Videogame 2",
+    "Lego Ninjago Movie Video Game",
+    "Lego Star Wars II: The Original Trilogy",
+    "Lego Star Wars: The Force Awakens",
+    "Lego Star Wars: The Skywalker Saga",
+    "Lego Star Wars: The Video Game",
+    "Lego The Hobbit",
+    "Lego Worlds",
+    "Legrand Legacy: Tale of the Fatebounds",
+    "Leisure Suit Larry: Magna Cum Laude",
+    "Leisure Suit Larry: Wet Dreams Don't Dry",
+    "Leisure Suit Larry: Wet Dreams Dry Twice",
+    "Lemmings",
+    "Lemnis Gate",
+    "Lemony Snicket's A Series of Unfortunate Events",
+    "Leo's Fortune: HD Edition",
+    "Let Them Come",
+    "Let's Sing 2020",
+    "Let's Sing 2021",
+    "Let's Sing Country",
+    "Let's Sing Queen",
+    "Lethal League",
+    "Lethal League Blaze",
+    "Letter Quest Remastered",
+    "Lichdom: Battlemage",
+    "Lichtspeer: Double Speer Edition",
+    "Life Goes On: Done to Death",
+    "Life Is Strange",
+    "Life Is Strange 2",
+    "Linda³",
+    "Link-a-Pix Deluxe",
+    "Links 2004",
+    "Lithium: Inmate 39",
+    "Little Big Workshop",
+    "Little Devil Inside",
+    "Little Misfortune",
+    "Little Nightmares",
+    "Little Nightmares II",
+    "Little Town Hero",
+    "Little Witch Academia: Chamber of Time",
+    "LittleBigPlanet 3",
+    "Livelock",
+    "Lock's Quest",
+    "LocoRoco 2 Remastered",
+    "LocoRoco Remastered",
+    "Lode Runner Legacy",
+    "Lode Runner: Ushina Wareta Meikyū",
+    "Lone Survivor: The Director's Cut",
+    "Lonely Mountains: Downhill",
+    "Loom",
+    "Loons: The Fight for Fame",
+    "Loot Rascals",
+    "Lord of Wars",
+    "Lords of Thunder - •Winds of ThunderJP",
+    "Lords of the Fallen",
+    "Lords of the Rising Sun - •Rising SunJP",
+    "Lornsword Winter Chronicle",
+    "Lost Castle",
+    "Lost Ember",
+    "Lost Grimoires 2: Shard of Mystery",
+    "Lost Grimoires: Stolen Kingdom",
+    "Lost Orbit: Terminal Velocity",
+    "Lost Ruins",
+    "Lost Sea",
+    "Lost Soul Aside",
+    "Lost Sphear",
+    "Lost Wing",
+    "Lost Words: Beyond the Page",
+    "Lotus Challenge",
+    "Louga II: The Ends of Shangrila",
+    "Lovecraft's Untold Stories",
+    "Lovers in a Dangerous Spacetime",
+    "MLB Slugfest 2003",
+    "MLB Slugfest 2004",
+    "MLB Slugfest 2006",
+    "MLB Slugfest: Loaded",
+    "MTV Music Generator 3: This is the Remix",
+    "MTX: Mototrax",
+    "MVP 06: NCAA Baseball",
+    "MVP Baseball 2003",
+    "MVP Baseball 2004",
+    "MVP Baseball 2005",
+    "MX 2002",
+    "MX Superfly",
+    "MX Unleashed",
+    "MX World Tour Featuring Jamie Little",
+    "MX vs. ATV Unleashed",
+    "Mace Griffin: Bounty Hunter",
+    "Mad Dash Racing",
+    "Mad Stalker: Full Metal Force",
+    "Madagascar",
+    "Madden NFL 06",
+    "Madden NFL 07",
+    "Madden NFL 08",
+    "Madden NFL 09",
+    "Madden NFL 2002",
+    "Madden NFL 2003",
+    "Madden NFL 2004",
+    "Madden NFL 2005",
+    "Madō King Granzort",
+    "Madō Monogatari I",
+    "Mafia: The City of Lost Heaven",
+    "Magatama",
+    "Magi Death Fight: Mahou Gakuen",
+    "Magic the Gathering: Battlegrounds",
+    "Magical Chase",
+    "Magical Dinosaur Tour - •Magical Saurus TourJP",
+    "Magicoal",
+    "Mahjong Clinic Special",
+    "Mahjong Gakuen - Tōma Sōhirō Tōjō",
+    "Mahjong Gakuen Mild",
+    "Mahjong Gokū Special",
+    "Mahjong Haōden: Kaiser's Quest",
+    "Mahjong Lemon Angel",
+    "Mahjong Shikaku Retsuden: Mahjong Wars",
+    "Mahjong Sword",
+    "Mahjong on the Beach",
+    "Maison Ikkoku",
+    "Major League Baseball 2K5",
+    "Major League Baseball 2K5: World Series Edition",
+    "Major League Baseball 2K6",
+    "Major League Baseball 2K7",
+    "Makai Hakkenden Shada",
+    "Makai Prince Dorabocchan",
+    "Malice",
+    "Mamono Hunter Yōko: Makai Kara no Tenkōsai",
+    "Mamono Hunter Yōko: Tooki Yobikoe",
+    "Manchester United Manager 2005",
+    "Manhunt",
+    "Maniac Pro Wrestling",
+    "Marc Ecko's Getting Up: Contents Under Pressure",
+    "Martial Champion",
+    "Marvel Nemesis: Rise of the Imperfects",
+    "Marvel vs. Capcom 2",
+    "Marvel: Ultimate Alliance",
+    "Mashed: Drive to Survive",
+    "Mashed: Fully Loaded",
+    "Master of Monsters",
+    "Mat Hoffman's Pro BMX 2",
+    "Max Payne",
+    "Max Payne 2: The Fall of Max Payne",
+    "Maximum Chase",
+    "MechAssault",
+    "MechAssault 2: Lone Wolf",
+    "Medal of Honor: European Assault",
+    "Medal of Honor: Frontline",
+    "Medal of Honor: Rising Sun",
+    "Mega Man Anniversary Collection",
+    "Megami Paradise",
+    "Meikyū no Elfine",
+    "Melbourne Cup Challenge - Frankie Dettori Racing (EUR)",
+    "Men of Valor",
+    "Mercenaries: Playground of Destruction",
+    "Metal Angel",
+    "Metal Angel 2",
+    "Metal Arms: Glitch in the System",
+    "Metal Dungeon",
+    "Metal Gear Solid 2: Substance",
+    "Metal Slug 3",
+    "Metal Slug 4",
+    "Metal Slug 5",
+    "Metal Stoker",
+    "Metal Wolf Chaos",
+    "MetamorJupiter",
+    "Miami Vice",
+    "Micro Machines",
+    "Midnight Club 3: DUB Edition",
+    "Midnight Club 3: DUB Edition Remix",
+    "Midnight Club II",
+    "Midtown Madness 3",
+    "Midway Arcade Treasures",
+    "Midway Arcade Treasures 2",
+    "Midway Arcade Treasures 3",
+    "Might & Magic",
+    "Might and Magic III: Isles of Terra",
+    "Mike Tyson Heavyweight Boxing",
+    "Military Madness - •NectarisJP",
+    "Minesweeper",
+    "Minority Report: Everybody Runs",
+    "Mirai Shōnen Conan",
+    "Mission Impossible: Operation Surma",
+    "Mitsubachi Gakuen",
+    "Mizubaku Daibouken",
+    "Mojo!",
+    "Momotaro Densetsu Gaiden Dai Ichi Shu",
+    "Momotarou Densetsu II",
+    "Momotarou Densetsu Turbo",
+    "Momotarō Katsugeki",
+    "Monopoly Party",
+    "Monster 4x4: World Circuit",
+    "Monster Garage",
+    "Monster Lair - •Wonder Boy III: Monster LairJP",
+    "Monster Maker: Yami no Ryūkishi",
+    "Monster Pro Wrestling",
+    "Moonlight Lady",
+    "Morita Shogi PC",
+    "Mortal Kombat: Armageddon",
+    "Mortal Kombat: Deadly Alliance",
+    "Mortal Kombat: Deception",
+    "Mortal Kombat: Shaolin Monks",
+    "Moto Roader",
+    "Moto Roader II",
+    "Moto Roader MC",
+    "MotoGP",
+    "MotoGP 2",
+    "MotoGP 3",
+    "Motocross Mania 3",
+    "Motteke Tamago",
+    "Mr Heli no Daibouken",
+    "Murakumo: Renegade Mech Pursuit",
+    "Muzzle Flash",
+    "Myst III: Exile",
+    "Myst IV: Revelation",
+    "Mystic Formula",
+    "Märchen Maze",
+    "N.U.D.E.@ Natural Ultimate Digital Experiment",
+    "NASCAR 06: Total Team Control",
+    "NASCAR 07",
+    "NASCAR 2005: Chase for the Cup",
+    "NASCAR Heat 2002",
+    "NASCAR Thunder 2002",
+    "NASCAR Thunder 2003",
+    "NASCAR Thunder 2004",
+    "NBA 2K2",
+    "NBA 2K3",
+    "NBA 2K6",
+    "NBA 2K7",
+    "NBA Ballers",
+    "NBA Ballers: Phenom",
+    "NBA Inside Drive 2002",
+    "NBA Inside Drive 2003",
+    "NBA Inside Drive 2004",
+    "NBA Jam",
+    "NBA Live 06",
+    "NBA Live 07",
+    "NBA Live 2002",
+    "NBA Live 2003",
+    "NBA Live 2004",
+    "NBA Live 2005",
+    "NBA Starting Five",
+    "NBA Street V3",
+    "NBA Street Vol. 2",
+    "NCAA College Basketball 2K3",
+    "NCAA College Football 2K3",
+    "NCAA Football 06",
+    "NCAA Football 07",
+    "NCAA Football 08",
+    "NCAA Football 2003",
+    "NCAA Football 2004",
+    "NCAA Football 2005",
+    "NCAA March Madness 06",
+    "NCAA March Madness 2004",
+    "NCAA March Madness 2005",
+    "NFL 2K2",
+    "NFL 2K3",
+    "NFL Blitz 2002",
+    "NFL Blitz 2003",
+    "NFL Blitz Pro",
+    "NFL Fever 2002",
+    "NFL Fever 2003",
+    "NFL Fever 2004",
+    "NFL Head Coach",
+    "NFL Street",
+    "NFL Street 2",
+    "NHK Taiga Drama Taiheki",
+    "NHL 06",
+    "NHL 07",
+    "NHL 2002",
+    "NHL 2003",
+    "NHL 2004",
+    "NHL 2005",
+    "NHL 2K3",
+    "NHL 2K6",
+    "NHL 2K7",
+    "NHL Hitz 20-02",
+    "NHL Hitz 20-03",
+    "NHL Hitz Pro",
+    "NHL Rivals 2004",
+    "Nakashima Tetsuya no Othello Seminar",
+    "Namco Museum",
+    "Namco Museum 50th Anniversary",
+    "Narc",
+    "Naritore: The Sugoroku '92",
+    "Naxat Open",
+    "Naxat Stadium",
+    "Nazo no Masquerade",
+    "Necros no Yōsai",
+    "Need for Speed: Carbon",
+    "Need for Speed: Hot Pursuit 2",
+    "Need for Speed: Most Wanted",
+    "Need for Speed: Underground",
+    "Need for Speed: Underground 2",
+    "Neighbours from Hell",
+    "Nekketsu Kōkō Dodgeball Bu: CD Soccer Hen",
+    "Nekketsu Kōkō Dodgeball Bu: PC Bangai Hen",
+    "Nekketsu Kōkō Dodgeball Bu: PC Soccer Hen",
+    "Nekketsu Legend Baseballer",
+    "Nemurenu Yoru Chiisaina Ohanashi",
+    "Neo Nectaris",
+    "Neutopia",
+    "Neutopia II",
+    "New Adventure Island - •Takahashi Meijin no Shin Bōken JimaJP",
+    "New Legends",
+    "Nexzr",
+    "Nexzr Special",
+    "Nickelodeon Party Blast",
+    "Night Creatures",
+    "NightCaster",
+    "NightCaster II: Equinox",
+    "Niko Niko Pun",
+    "Ninja Gaiden",
+    "Ninja Gaiden Black",
+    "Ninja Ryūkenden",
+    "Ninja Spirit - •Saigo no Nindō: Ninja SpiritJP",
+    "Ninja Warriors",
+    "Nishimura Kyōtaro Mystery: Hokutosei no Onna",
+    "No-Ri-Ko",
+    "Nobunaga no Yabou: Ranseiki",
+    "Nobunaga no Yabō: Bushō Fūunroku",
+    "Nobunaga no Yabō: Zenkokuban",
+    "Obocchamakun",
+    "ObsCure",
+    "Oddworld: Munch's Oddysee",
+    "Oddworld: Stranger's Wrath",
+    "Open Season",
+    "Operation Flashpoint: Elite",
+    "Operation Wolf",
+    "Ordyne",
+    "Otogi 2: Immortal Warriors",
+    "Otogi: Myth of Demons",
+    "Out Live",
+    "Out Run",
+    "OutRun 2",
+    "OutRun 2006: Coast 2 Coast",
+    "Outlaw Golf",
+    "Outlaw Golf 2",
+    "Outlaw Golf: 9 More Holes of X-Mas",
+    "Outlaw Golf: Holiday Golf (9 Holes of X-Mas)",
+    "Outlaw Tennis",
+    "Outlaw Volleyball",
+    "Outlaw Volleyball: Red Hot",
+    "Over the Hedge",
+    "OverRide",
+    "P-47",
+    "Pac-Land",
+    "Pac-Man World 2",
+    "Pac-Man World 3",
+    "Pachio-kun 3: Pachislot & Pachinko",
+    "Pachio-kun: Jūban Shōbu",
+    "Pachio-kun: Maboroshi no Densetsu",
+    "Pachio-kun: Warau Uchū",
+    "Painkiller: Hell Wars",
+    "Panic Bomber",
+    "Panzer Dragoon Orta",
+    "Panzer Elite Action: Fields of Glory",
+    "Parasol Stars",
+    "Pariah",
+    "Parodius Da!",
+    "Pastel Lime",
+    "Peter Jackson's King Kong",
+    "Petit Copter",
+    "Phantasy Star Online Episode I & II",
+    "Phantom Crash",
+    "Phantom Dust",
+    "Pilot Down: Behind Enemy Lines",
+    "Pinball Hall of Fame",
+    "Pirates of the Caribbean",
+    "Pirates: Legend of the Black Buccaneer",
+    "Pirates: The Legend of Black Kat",
+    "Pitfall: The Lost Expedition",
+    "Playboy: The Mansion",
+    "Plus Plum 2",
+    "PocketBike Racer",
+    "Police Connection",
+    "Pool Shark 2",
+    "Pop'n Magic",
+    "Popful Mail",
+    "Populous",
+    "Populous: The Promised Lands",
+    "Power Drift",
+    "Power Eleven",
+    "Power Gate",
+    "Power Golf",
+    "Power Golf 2",
+    "Power League '93",
+    "Power League 4",
+    "Power League II",
+    "Power League III",
+    "Power League V",
+    "Power Tennis",
+    "Powerdrome",
+    "Predator: Concrete Jungle",
+    "Prince of Persia",
+    "Prince of Persia: The Sands of Time",
+    "Prince of Persia: The Two Thrones",
+    "Prince of Persia: Warrior Within",
+    "Princess Maker 1",
+    "Princess Maker 2",
+    "Princess Minerva",
+    "Prisoner of War",
+    "Private Eyedol",
+    "Pro Cast Sports Fishing - Lakemasters: Bass Fishing Game (JP)",
+    "Pro Fishing Challenge",
+    "Pro Race Driver",
+    "Pro Yakyuu World Stadium",
+    "Pro Yakyuu World Stadium '91",
+    "ProStroke Golf: World Tour 2007",
+    "Project Gotham Racing",
+    "Project Gotham Racing 2",
+    "Project Snowblind",
+    "Psi-Ops: The Mindgate Conspiracy",
+    "Psychic Detective Series Vol. 3: Aýa",
+    "Psychic Detective Vol. 4: Orgel",
+    "Psychic Storm",
+    "Psycho Chaser",
+    "Psychonauts",
+    "Psychosis - •ParanoiaJP",
+    "Psyvariar 2",
+    "Pulse Racer",
+    "Pump It Up: Exceed",
+    "Pure Pinball",
+    "Puyo Pop: Fever",
+    "Puyo Puyo CD",
+    "Puyo Puyo CD Tsū",
+    "Puzzle Boy",
+    "Puzznic",
+    "Quantum Redshift",
+    "Quiz Avenue",
+    "Quiz Avenue 3",
+    "Quiz Avenue II",
+    "Quiz Caravan Cult Q",
+    "Quiz Marugoto The World",
+    "Quiz Marugoto The World: Time Machine ni Onegai!",
+    "Quiz Nobunaga no Yabō",
+    "Quiz de Gakuensai",
+    "Quiz no Hoshi",
+    "R-Type - •R-Type I / R-Type IIJP",
+    "R-Type: Complete CD",
+    "R: Racing Evolution",
+    "RLH: Run Like Hell",
+    "ROM ROM Stadium",
+    "Rabio Lepus Special",
+    "Racing Spirits",
+    "Raiden",
+    "Rainbow Islands",
+    "RalliSport Challenge 2",
+    "Rallisport Challenge",
+    "Rally Fusion: Race of Champions",
+    "Ranma ½",
+    "Ranma ½: Datō, Ganso Musabetsu Kakuto Ryū",
+    "Ranma ½: Toraware no Hayanome",
+    "Rapala Pro Fishing",
+    "Rastan Saga II",
+    "Ratatouille",
+    "Rayman 3: Hoodlum Havoc",
+    "Rayman Arena",
+    "Rayxanber II",
+    "Rayxanber III",
+    "Raze's Hell",
+    "Real World Golf",
+    "Record of Lodoss War",
+    "Records of Lodoss War II",
+    "Red Dead Revolver",
+    "Red Faction II",
+    "Red Ninja: End of Honor",
+    "RedCard 20-03",
+    "Reign of Fire",
+    "Renny Blaster",
+    "Rent-A-Hero No. 1",
+    "Reservoir Dogs",
+    "Return to Castle Wolfenstein: Tides of War",
+    "Richard Burns Rally",
+    "Riot Zone - •Crest of WolfJP",
+    "Road Spirits",
+    "RoadKill",
+    "Robin Hood: Defender of the Crown",
+    "RoboCop",
+    "Robot Wars: Extreme Destruction",
+    "Robotech: Battlecry",
+    "Robotech: Invasion",
+    "Robots",
+    "Rock-On",
+    "Rocky",
+    "Rocky: Legends",
+    "Rogue Ops",
+    "Rogue Trooper",
+    "RollerCoaster Tycoon",
+    "Rolling",
+    "Room Zoom",
+    "Rugby 06",
+    "Rugby 2005",
+    "Rugby Challenge 2006",
+    "Rugby League",
+    "Rugby League 2",
+    "Ruin: Kami no Isan",
+    "Ryūkyū",
+    "Ryūkō no Ken",
+    "S.C.I.",
+    "SD Senyō Aldynes",
+    "SSX 3",
+    "SSX On Tour",
+    "SSX Tricky",
+    "SVC Chaos: SNK vs. Capcom",
+    "SWAT: Global Strike Team",
+    "SX Superstar",
+    "Sadakichi Sebun: Hideyoshi no Ougon",
+    "Saint Dragon",
+    "Salamander",
+    "Samurai Ghost - •Genpei Toumaden: KannoniJP",
+    "Samurai Shodown V",
+    "Samurai Warriors",
+    "Sangokushi III",
+    "Sangokushi: Eiketsu Tenka ni Nozomu",
+    "Scaler",
+    "Scarface: The World Is Yours",
+    "Scooby-Doo! Mystery Mayhem",
+    "Scooby-Doo! Night of 100 Frights",
+    "Scooby-Doo! Unmasked",
+    "SeaBlade",
+    "SeaWorld: Shamu's Deep Sea Adventures",
+    "Second Sight",
+    "Secret Weapons Over Normandy",
+    "Sega GT 2002",
+    "Sega GT Online",
+    "Sega Soccer Slam",
+    "Seirei Senshi Spriggan",
+    "Seiryū Densetsu Monbit",
+    "Seisenshi Denshō",
+    "Seiya Monogatari",
+    "Sekigahara",
+    "Sengoku Kantō Sangokushi",
+    "Sengoku Mahjong",
+    "Sensible Soccer 2006",
+    "Sentou Yousei Yukikaze: Yousei no Mau Sora",
+    "Serious Sam",
+    "Serious Sam II",
+    "Sexy Idol Mahjong",
+    "Sexy Idol Mahjong 2: Yakyūken no Uta",
+    "Sexy Idol Mahjong: Mahjong Fashion Monogatari",
+    "Shadow Ops: Red Mercury",
+    "Shadow of Memories",
+    "Shadow of the Beast",
+    "Shadow the Hedgehog",
+    "Shanghai",
+    "Shanghai II",
+    "Shanghai III: Dragon's Eye",
+    "Shape Shifter - •Shapeshift: Makai Eiyū DenJP",
+    "Shark Tale",
+    "Shattered Union",
+    "Shellshock: Nam '67",
+    "Shenmue II",
+    "Sherlock Holmes Consulting Detective",
+    "Sherlock Holmes: Consulting Detective Vol. II",
+    "Shikigami no Shiro",
+    "Shikigami no Shiro Evolution Blue",
+    "Shikigami no Shiro Evolution Red",
+    "Shikigami no Shiro II",
+    "Shin Megami Tensei",
+    "Shin Megami Tensei: Nine",
+    "Shin Onryō Senki",
+    "Shin Sangokushi: Tenka wa Waga ni",
+    "Shinchou Mahjong (Nobunaga Mahjong)",
+    "Shinobi",
+    "Shiryō Sensen",
+    "Shockman - •Kaizou Chōjin Shubibinman 2: Atanaru TekiJP",
+    "Shogi Shodan Icchokusen",
+    "Shogi Shoshinsha Muyou",
+    "Showdown: Legends of Wrestling",
+    "Shrek",
+    "Shrek 2",
+    "Shrek Super Party",
+    "Shrek SuperSlam",
+    "Shōgi Database Kiyū",
+    "Sid Meier's Pirates!",
+    "Side Arms Hyper Dyne - •Hyper Dyne Side ArmsJP",
+    "Side Arms Special",
+    "Silent Debuggers",
+    "Silent Hill 2",
+    "Silent Hill 4: The Room",
+    "Silent Scope Complete",
+    "SimEarth",
+    "Sindibad: Chitei No Dai Makyu",
+    "Sinistron - •Violent SoldierJP",
+    "Ski Racing 2005",
+    "Ski Racing 2006",
+    "Skweek",
+    "Slam Tennis",
+    "Slime World",
+    "Slot Gambler",
+    "Smashing Drive",
+    "Snatcher CD-ROMantic",
+    "Sneak King",
+    "Sneakers",
+    "Sniper Elite",
+    "Sol Bianca",
+    "Sol Moonarge",
+    "Soldier Blade",
+    "Soldier of Fortune II: Double Helix",
+    "Solid Force",
+    "Somer Assault - •MesopotamiaJP",
+    "Son Son II",
+    "Sonic Heroes",
+    "Sonic Mega Collection Plus",
+    "Sonic Riders",
+    "Sonic Spike - •World Beach VolleyJP",
+    "Sorcerian",
+    "Sotsugyō II",
+    "Sotsugyō Shashin: Miki",
+    "Sotsugyō: Graduation",
+    "Soul Calibur II",
+    "Space Harrier",
+    "Space Invaders: Fukkatsu no Hi",
+    "Space Invaders: The Original Game",
+    "Spartan: Total Warrior",
+    "Spawn: Armageddon",
+    "Special Forces: Nemesis Strike - Counter Terrorist Special Forces: Fire for Effect (EUR)",
+    "Speed Kings",
+    "Sphinx and the Cursed Mummy",
+    "Spider-Man",
+    "Spider-Man 2",
+    "Spikeout: Battle Street",
+    "Spin Pair",
+    "Spiral Wave",
+    "Splash Lake",
+    "Splashdown",
+    "Splat Magazine Renegade Paintball",
+    "Splatterhouse",
+    "SpongeBob SquarePants: Battle for Bikini Bottom",
+    "SpongeBob SquarePants: Lights, Camera, Pants!",
+    "Spriggan Mk. II",
+    "Spy Hunter",
+    "Spy Hunter 2",
+    "Spy Hunter: Nowhere to Run",
+    "Spy vs. Spy",
+    "Spyro: A Hero's Tail",
+    "Stacked with Daniel Negreanu",
+    "Stake: Fortune Fighters",
+    "Star Breaker",
+    "Star Mobile",
+    "Star Parodier",
+    "Star Trek: Shattered Universe",
+    "Star Wars: Battlefront",
+    "Star Wars: Battlefront II",
+    "Star Wars: Episode III: Revenge of the Sith",
+    "Star Wars: Jedi Knight II: Jedi Outcast",
+    "Star Wars: Jedi Knight: Jedi Academy",
+    "Star Wars: Jedi Starfighter",
+    "Star Wars: Knights of the Old Republic",
+    "Star Wars: Knights of the Old Republic II: The Sith Lords",
+    "Star Wars: Obi-Wan",
+    "Star Wars: Republic Commando",
+    "Star Wars: Starfighter",
+    "Star Wars: The Clone Wars",
+    "Starsky & Hutch",
+    "Startling Odyssey",
+    "Startling Odyssey II",
+    "State of Emergency",
+    "Steam-Heart's",
+    "Steel Battalion",
+    "Steel Battalion: Line of Contact",
+    "Still Life",
+    "Stolen",
+    "Stratego",
+    "Street Fighter Anniversary Collection",
+    "Street Fighter II",
+    "Street Hoops",
+    "Street Racing Syndicate",
+    "Strider Hiryū",
+    "Strike Force Bowling",
+    "Stubbs the Zombie in Rebel Without a Pulse",
+    "Sudeki",
+    "Super Air Zonk - •CD-ROM² DenjinJP",
+    "Super Albatross",
+    "Super Bubble Pop",
+    "Super CD-ROM² Taiken Soft Shū",
+    "Super Daisenryaku",
+    "Super Darius",
+    "Super Darius II",
+    "Super Mahjong Taikai",
+    "Super Metal Crusher",
+    "Super Momotaro Dentetsu",
+    "Super Momotarou Dentetsu II",
+    "Super Monkey Ball Deluxe",
+    "Super Raiden",
+    "Super Real Mahjong P II/III Custom",
+    "Super Real Mahjong P.V Custom",
+    "Super Real Mahjong PIV",
+    "Super Real Mahjong Special",
+    "Super Schwarzchild",
+    "Super Schwarzschild 2",
+    "Super Star Soldier",
+    "Super Volleyball",
+    "Superman Returns",
+    "Superman: The Man of Steel",
+    "Susano-ou Densetsu",
+    "Sword Master",
+    "Syberia",
+    "Syberia II",
+    "Syd Mead's Terraforming - •TerraformingJP",
+    "Sylphia",
+    "TATSUJIN",
+    "TOCA Race Driver 2: The Ultimate Racing Simulator",
+    "TOCA Race Driver 3",
+    "TV Sports Basketball",
+    "TV Sports Hockey - •TV Sports Ice HockeyJP",
+    "TV Sports: Football",
+    "Taidaima Yūsha Boshūchū",
+    "Taiheiki",
+    "Taito Legends",
+    "Taito Legends 2",
+    "Tak 2: The Staff of Dreams",
+    "Tak: The Great Juju Challenge",
+    "Takahashi Akiko no Mahjong Seminar",
+    "Takeda Shingen",
+    "Takin' it to the Hoop - •USA Pro BasketballJP",
+    "TaleSpin",
+    "Tanjō: Debut",
+    "Tao Feng: Fist of the Lotus",
+    "Tatsu no Ko Fighter",
+    "Taz: Wanted",
+    "Tecmo Classic Arcade",
+    "Tecmo World Cup: Super Soccer",
+    "Teen Titans",
+    "Teenage Mutant Ninja Turtles",
+    "Teenage Mutant Ninja Turtles 2: Battle Nexus",
+    "Teenage Mutant Ninja Turtles 3: Mutant Nightmare",
+    "Teenage Mutant Ninja Turtles: Mutant Melee",
+    "Tenchi Muyō! Ryōōki",
+    "Tenchi wo Kurau",
+    "Tenchu: Return from Darkness",
+    "Tenerezza",
+    "Tengai Makyō II: Manjimaru",
+    "Tengai Makyō: Deden no Den",
+    "Tengai Makyō: Fū-un Kabuki Den",
+    "Tengai Makyō: ZIRIA",
+    "Tennis Masters Series 2003",
+    "Tenshi no Uta",
+    "Tenshi no Uta II",
+    "Terminator 3: Rise of the Machines",
+    "Terminator 3: The Redemption",
+    "Terra Cresta II",
+    "Test Drive",
+    "Test Drive Off-Road Wide Open",
+    "Test Drive: Eve of Destruction",
+    "Tetris Worlds",
+    "Tetris Worlds (Online Edition)",
+    "The 25th Ward: The Silver Case",
+    "The Adventure Pals",
+    "The Alliance Alive HD Remastered",
+    "The Alto Collection",
+    "The Amazing American Circus",
+    "The Amazing Spider-Man 2",
+    "The Angry Video Game Nerd I & II Deluxe",
+    "The Aquatic Adventure of the Last Human",
+    "The Assembly",
+    "The Atlas",
+    "The Banner Saga",
+    "The Banner Saga 2",
+    "The Banner Saga 3",
+    "The Bard's Tale",
+    "The Bard's Tale IV: Director's Cut",
+    "The Bard's Tale: Remastered and Resnarkled",
+    "The Baseball 2002: Battle Ball Park Sengen",
+    "The Bible Game",
+    "The Binding of Isaac: Rebirth",
+    "The Binding of Isaac: Repentance",
+    "The Blackout Club",
+    "The Book of Unwritten Tales 2",
+    "The Bridge",
+    "The Bug Butcher",
+    "The Caligula Effect 2",
+    "The Caligula Effect: Overdose",
+    "The Castle Game",
+    "The Chronicles of Narnia: The Lion, the Witch and the Wardrobe",
+    "The Chronicles of Riddick: Escape from Butcher Bay",
+    "The Church in the Darkness",
+    "The Coma 2: Vicious Sisters",
+    "The Coma: Recut",
+    "The Count Lucanor",
+    "The Crew",
+    "The Crew 2",
+    "The Da Vinci Code",
+    "The Dark Pictures Anthology: House of Ashes",
+    "The Dark Pictures Anthology: Little Hope",
+    "The Dark Pictures Anthology: Man of Medan",
+    "The Deadly Tower of Monsters",
+    "The Disney Afternoon Collection",
+    "The Dukes of Hazzard: Return of the General Lee",
+    "The Dwarves",
+    "The Elder Scrolls III: Morrowind",
+    "The Elder Scrolls Online",
+    "The Elder Scrolls V: Skyrim – Special Edition",
+    "The End Is Nigh",
+    "The Escapists",
+    "The Escapists 2",
+    "The Escapists: The Walking Dead",
+    "The Evil Within",
+    "The Evil Within 2",
+    "The Fairly OddParents: Breakin' Da Rules",
+    "The Fall",
+    "The Fall Part 2: Unbound",
+    "The Final Station",
+    "The Flame in the Flood",
+    "The Forest",
+    "The Gardens Between",
+    "The Godfather: The Game",
+    "The Golf Club",
+    "The Golf Club 2",
+    "The Golf Club 2019 featuring PGA Tour",
+    "The Good Life",
+    "The Great Ace Attorney Chronicles",
+    "The Great Escape",
+    "The Guy Game",
+    "The Haunted Mansion",
+    "The Hobbit",
+    "The Hong Kong Massacre",
+    "The House in Fata Morgana: Dream of the Revenants Edition",
+    "The House of the Dead III",
+    "The Huntsman: Winter's Curse",
+    "The Hustle: Detroit Streets",
+    "The Idolmaster Platinum Stars",
+    "The Idolmaster: Starlit Season",
+    "The Idolmaster: Stella Stage",
+    "The Incredible Adventures of Van Helsing",
+    "The Incredible Adventures of Van Helsing II",
+    "The Incredible Adventures of Van Helsing III",
+    "The Incredible Hulk: Ultimate Destruction",
+    "The Incredibles - Mr. Incredible (JP)",
+    "The Incredibles: Rise of the Underminer",
+    "The Inner Friend",
+    "The Inner World",
+    "The Inpatient",
+    "The Invisible Hours",
+    "The Italian Job",
+    "The Jackbox Party Pack",
+    "The Jackbox Party Pack 2",
+    "The King of Fighters '97 Global Match",
+    "The King of Fighters 2002",
+    "The King of Fighters 2002: Unlimited Match",
+    "The King of Fighters 2003",
+    "The King of Fighters Neowave",
+    "The King of Fighters XIV",
+    "The King of Fighters: Maximum Impact",
+    "The Language of Love",
+    "The Last Blade 2",
+    "The Last Campfire",
+    "The Last Door",
+    "The Last Guardian",
+    "The Last Kids on Earth and the Staff of Doom",
+    "The Last Remnant Remastered",
+    "The Last Tinker: City of Colors",
+    "The Last of Us Part II",
+    "The Last of Us Remastered",
+    "The Last of Us: Left Behind",
+    "The Legend of Heroes: Ao no Kiseki",
+    "The Legend of Heroes: Hajimari no Kiseki",
+    "The Legend of Heroes: Trails of Cold Steel",
+    "The Legend of Heroes: Trails of Cold Steel II",
+    "The Legend of Heroes: Trails of Cold Steel III",
+    "The Legend of Heroes: Trails of Cold Steel IV",
+    "The Legend of Heroes: Zero no Kiseki",
+    "The Legend of Korra",
+    "The Legend of Spyro: A New Beginning",
+    "The Little Acre",
+    "The Long Dark",
+    "The Lord of the Rings: Adventure Card Game",
+    "The Lord of the Rings: The Fellowship of the Ring",
+    "The Lord of the Rings: The Return of the King",
+    "The Lord of the Rings: The Third Age",
+    "The Lord of the Rings: The Two Towers",
+    "The Lost Child",
+    "The Manhole",
+    "The Matrix: Path of Neo",
+    "The NewZealand Story",
+    "The Pro Yakyū",
+    "The Pro Yakyū Super",
+    "The Pro Yakyū Super '94",
+    "The Punisher",
+    "The Simpsons: Hit & Run",
+    "The Simpsons: Road Rage",
+    "The Sims",
+    "The Sims 2",
+    "The Sims Bustin' Out",
+    "The SpongeBob SquarePants Movie Game",
+    "The Suffering",
+    "The Suffering: Ties That Bind",
+    "The TV Show",
+    "The Terminator: Dawn of Fate",
+    "The Thing",
+    "The Tower of Druaga",
+    "The Urbz: Sims in the City",
+    "The Warriors",
+    "The Wild Rings",
+    "Thief: Deadly Shadows",
+    "Thousand Land",
+    "Thrillville",
+    "Thunder Blade",
+    "Tiger Road - •Tora he no MichiJP",
+    "Tiger Woods PGA Tour 06",
+    "Tiger Woods PGA Tour 07",
+    "Tiger Woods PGA Tour 2003",
+    "Tiger Woods PGA Tour 2004",
+    "Tiger Woods PGA Tour 2005",
+    "Tim Burton's The Nightmare Before Christmas: Oogie's Revenge",
+    "Time Cruise - •Time Cruise IIJP",
+    "TimeSplitters 2",
+    "TimeSplitters: Future Perfect",
+    "Timeball - •BlodiaJP",
+    "Titan",
+    "ToeJam & Earl III: Mission to Earth",
+    "Toilet Kids",
+    "Tokimeki Memorial",
+    "Tom Clancy's Ghost Recon",
+    "Tom Clancy's Ghost Recon 2",
+    "Tom Clancy's Ghost Recon 2: Summit Strike",
+    "Tom Clancy's Ghost Recon: Advanced Warfighter",
+    "Tom Clancy's Ghost Recon: Island Thunder",
+    "Tom Clancy's Rainbow Six 3",
+    "Tom Clancy's Rainbow Six 3: Black Arrow",
+    "Tom Clancy's Rainbow Six: Critical Hour",
+    "Tom Clancy's Rainbow Six: Lockdown",
+    "Tom Clancy's Splinter Cell",
+    "Tom Clancy's Splinter Cell: Chaos Theory",
+    "Tom Clancy's Splinter Cell: Double Agent",
+    "Tom Clancy's Splinter Cell: Pandora Tomorrow",
+    "Tom and Jerry in War of the Whiskers",
+    "Tomb Raider: Legend",
+    "Tony Hawk's American Wasteland",
+    "Tony Hawk's Pro Skater 2x",
+    "Tony Hawk's Pro Skater 3",
+    "Tony Hawk's Pro Skater 4",
+    "Tony Hawk's Project 8",
+    "Tony Hawk's Underground",
+    "Tony Hawk's Underground 2",
+    "Top Gear RPM Tuning - RPM Tuning (EUR)",
+    "Top Spin Tennis",
+    "Top o Nerae! GunBuster Vol. 1",
+    "Top o Nerae! GunBuster Vol. 2",
+    "Torino 2006",
+    "Tork: Prehistoric Punk",
+    "Toshi Tensō Keikaku Eternal City",
+    "Total Club Manager 2004",
+    "Total Club Manager 2005",
+    "Total Immersion Racing",
+    "Total Overdose: A Gunslinger's Tale in Mexico",
+    "Totaled!",
+    "Touge R",
+    "Tour de France",
+    "Toxic Grind",
+    "Toy Shop Boys",
+    "TransWorld Snowboarding",
+    "TransWorld Surf",
+    "Travel Epuru",
+    "Travelers!: Densetsu o Buttobase",
+    "Triangle Again",
+    "Triangle Again 2",
+    "Tricky Kick - •TrickyJP",
+    "Trigger Man",
+    "Triple Play 2002",
+    "Trivial Pursuit Unhinged",
+    "Tron 2.0 Killer App",
+    "True Crime: New York City",
+    "True Crime: Streets of LA",
+    "Tsuppari Oozumou Heisei Ban",
+    "Tsuru Teruto no Jissen Kabushiki Bi-Game",
+    "Turok: Evolution",
+    "Turrican",
+    "Ty the Tasmanian Tiger",
+    "Ty the Tasmanian Tiger 2: Bush Rescue",
+    "Ty the Tasmanian Tiger 3: Night of the Quinkan",
+    "UEFA Champions League 2004-2005",
+    "UEFA Euro 2004",
+    "UFC: Tapout",
+    "UFC: Tapout 2",
+    "Uchū Senkan Yamato",
+    "Ultimate Beach Soccer",
+    "Ultimate Pro Pinball",
+    "Ultimate Spider-Man",
+    "Ultra Box No. 2",
+    "Ultra Box No. 3",
+    "Ultra Box No. 4",
+    "Ultra Box No. 5",
+    "Ultra Box No. 6",
+    "Ultra Box Premiere Issue",
+    "Ultra Bust-a-Move - Ultra Puzzle Bobble (JP)",
+    "Umezawa Yukari no Igo Seminar",
+    "Unreal Championship",
+    "Unreal Championship 2: The Liandri Conflict",
+    "Unreal II: The Awakening",
+    "Urban Chaos: Riot Response",
+    "Urusei Yatsura: Stay With You",
+    "V-Rally 3",
+    "Valis II",
+    "Valis III",
+    "Valis IV",
+    "Valis: The Fantasm Soldier",
+    "Valkyrie no Densetsu",
+    "Van Helsing",
+    "Vanilla Syndrome",
+    "Vasteel",
+    "Vasteel 2",
+    "Veigues Tactical Gladiator - •VeiguesJP",
+    "Vexx",
+    "Victory Run",
+    "Vietcong: Purple Haze",
+    "Vigilante",
+    "Virgin Dream",
+    "Virtual Pool: Tournament Edition",
+    "Volfied",
+    "Volvo: Drive For Life",
+    "Voodoo Vince",
+    "WTA Tour Tennis",
+    "WWE Raw 2",
+    "WWE WrestleMania 21",
+    "WWF RAW",
+    "Wai Wai Mahjong",
+    "Wakeboarding Unleashed",
+    "Wallaby!!",
+    "Wallace & Gromit in Project Zoo",
+    "Wallace & Gromit: The Curse of the Were-Rabbit",
+    "Warpath",
+    "Whacked!",
+    "Where in the World is Carmen Sandiego?",
+    "Whiplash",
+    "Whiteout",
+    "WinBack 2: Project Poseidon",
+    "Wings of War",
+    "Winning Shot",
+    "Without Warning",
+    "Wizardry I & II",
+    "Wizardry III & IV",
+    "Wizardry V",
+    "Wonder Momo",
+    "World Championship Poker",
+    "World Championship Poker 2: Featuring Howard Lederer",
+    "World Championship Pool 2004",
+    "World Championship Rugby",
+    "World Championship Snooker 2003",
+    "World Championship Snooker 2004",
+    "World Circuit",
+    "World Class Baseball - •Power LeagueJP",
+    "World Court Tennis - •Pro Tennis: World CourtJP",
+    "World Heroes 2",
+    "World Jockey",
+    "World Poker Tour",
+    "World Racing",
+    "World Racing 2",
+    "World Series Baseball 2K2",
+    "World Series Baseball 2K3",
+    "World Series of Poker",
+    "World Snooker Championship 2005",
+    "World Soccer Winning Eleven 8 International - Pro Evolution Soccer 4 (PAL) - Winning Eleven 8 (JP)",
+    "World Soccer Winning Eleven 9 - Pro Evolution Soccer 5 (PAL) - Winning Eleven 9 (JP)",
+    "World Sports Competition - •Power SportsJP",
+    "World War II Combat: Iwo Jima",
+    "World War II Combat: Road to Berlin",
+    "Worms 3D",
+    "Worms 4: Mayhem",
+    "Worms Forts: Under Siege",
+    "Wrath Unleashed",
+    "Wreckless: The Yakuza Missions - Double-S.T.E.A.L. (JP)",
+    "Wrestle Angels: Double Impact",
+    "X-Men Legends",
+    "X-Men Legends II: Rise of Apocalypse",
+    "X-Men: Next Dimension",
+    "X-Men: The Official Game",
+    "X2: Wolverine's Revenge",
+    "XGRA: Extreme-G Racing Association",
+    "XIII",
+    "Xak I & II",
+    "Xak III: The Eternal Recurrence",
+    "Xevious: Fardraut Saga",
+    "Xiaolin Showdown",
+    "Xyanide",
+    "Yager",
+    "Yamamura Misa Suspense: Kizenka Kyō Ezara Satsujin Jiken",
+    "Yami no Ketsuzoku",
+    "Yawara!",
+    "Yawara! 2",
+    "Yetisports Arctic Adventures",
+    "Yo, Bro",
+    "Yokai Dochuki",
+    "Yonenaga Kunio no Shougi Seminar",
+    "Yourself!Fitness",
+    "Ys III: Wanderers from Ys",
+    "Ys IV: The Dawn of Ys",
+    "Ys: Book I & II - •Ys I & IIJP",
+    "Yu-Gi-Oh! The Dawn of Destiny",
+    "Yū Yū Hakusho: Yami Shōbu!! Ankoku Bujutsu Kai",
+    "Yūyū Jinsei",
+    "Zan: Kagerō no Toki",
+    "Zapper: One Wicked Cricket",
+    "Zathura",
+    "Zero Wing",
+    "Zero4 Champ",
+    "Zero4 Champ II",
+    "ZillerNet#Funtime",
+    "Zipang",
+    "eBaseball Powerful Pro Yakyuu 2020",
+    "eFootball PES 2021 Season Update",
+    "eFootball Pro Evolution Soccer 2020",
+    "iO",
+    "Æon Flux",
+});
+
+const auto videoGameGenres = std::to_array<std::string_view>({
+    "Action",
+    "Adventure",
+    "Battle royale",
+    "Dating sim",
+    "Endless runner",
+    "Fighting",
+    "First-person shooter",
+    "Hack and slash",
+    "Horror",
+    "Massively multiplayer online",
+    "Music",
+    "Platformer",
+    "Puzzle",
+    "Racing",
+    "Real-time strategy",
+    "Roguelike",
+    "Role-playing game",
+    "Sandbox",
+    "Science fiction game",
+    "Shooter",
+    "Sim racing",
+    "Soulslike",
+    "Sports",
+    "Stealth",
+    "Strategy",
+    "Survival",
+    "Third-person shooter",
+    "Tower defense",
+});
+
+const auto platforms = std::to_array<std::string_view>({
+    "Android",       "Linux",  "Nintendo Switch", "Oculus Quest", "PC",
+    "Playstation 5", "Stadia", "Xbox Series X",   "iOS",
+});
+
+const auto studioNames = std::to_array<std::string_view>({
+    "0verflow",
+    "1st Playable Productions",
+    "2K Czech",
+    "989 Studios",
+    "ACE Team",
+    "AQ Interactive",
+    "Access Games",
+    "Acclaim Entertainment",
+    "Accolade",
+    "Acheron Design",
+    "Acquire",
+    "Active Gaming Media",
+    "Activision Blizzard",
+    "Adventure Soft",
+    "Akella",
+    "Aki Corporation",
+    "Alfa System",
+    "Ancient",
+    "Anino Games",
+    "Arc System Works",
+    "ArenaNet",
+    "Arika",
+    "Arkane Studios",
+    "Arkedo Studio",
+    "Artdink",
+    "ArtePiazza",
+    "Artificial Studios",
+    "Artoon",
+    "Ascaron",
+    "Asobo Studio",
+    "Aspect",
+    "Aspyr Media",
+    "Atari",
+    "Atlus",
+    "Atomic Planet Entertainment",
+    "Attic Entertainment Software",
+    "Avalanche Software",
+    "Avalanche Studios",
+    "Aventurine SA",
+    "BSure Interactive",
+    "Babaroga",
+    "Backbone Entertainment",
+    "Banpresto",
+    "BattleGoat Studios",
+    "Beenox",
+    "Behemoth",
+    "Bethesda Softworks",
+    "Big Blue Bubble",
+    "Big Huge Games",
+    "Binary Hammer",
+    "BioWare",
+    "Bizarre Creations",
+    "Black Rock Studio",
+    "Blitz Games Studios",
+    "Blizzard Entertainment",
+    "Blue Byte Software",
+    "Blue Fang Games",
+    "Bohemia Interactive",
+    "BreakAway Games",
+    "Brownie Brown",
+    "Brøderbund",
+    "Bugbear Entertainment",
+    "Buka Entertainment",
+    "Bullfrog Productions",
+    "Bungie Studios",
+    "CCP Games",
+    "CD Projekt RED",
+    "Capcom",
+    "Cave",
+    "Cavia",
+    "Centuri",
+    "Chunsoft",
+    "Cinemaware",
+    "Cing",
+    "Clap Hanz",
+    "Climax Entertainment",
+    "Climax Studios",
+    "Coded Illusions",
+    "Codemasters",
+    "Coktel Vision",
+    "ColdWood Interactive",
+    "Compile Heart",
+    "Core Design",
+    "Crafts & Meister",
+    "Creat Studios",
+    "Creative Assembly",
+    "Criterion Games",
+    "Cryptic Studios",
+    "Crystal Dynamics",
+    "Crytek",
+    "Cyanide",
+    "CyberConnect2",
+    "Cyberlore Studios",
+    "Day 1 Studios",
+    "Deadline Games",
+    "Deck13",
+    "Deep Silver",
+    "Demiurge Studios",
+    "Digital Illusions CE",
+    "Dimps",
+    "Disney Interactive Studios",
+    "Double Fine Productions",
+    "Double Helix Games",
+    "Dynamite Idea",
+    "EDGE Games",
+    "Egosoft",
+    "Eidos Interactive",
+    "Electronic Arts/EA Games",
+    "Engine Software",
+    "Epic Games",
+    "Epicenter Studios",
+    "Epyx",
+    "Etranges Libellules",
+    "Eugen Systems",
+    "Eurocom",
+    "Evolution Studios",
+    "FUN Labs",
+    "FarSight Studios",
+    "Fatshark",
+    "Firaxis Games",
+    "Firefly Studios",
+    "First Star Software",
+    "Flagship Games",
+    "Flying Lab Software",
+    "Foundation 9 Entertainment",
+    "Free Radical Design",
+    "Frictional Games",
+    "From Software",
+    "Frontier Developments",
+    "Frozenbyte",
+    "Funcom",
+    "Futuremark",
+    "GSC Game World",
+    "Game Arts",
+    "Game Freak",
+    "GameHouse",
+    "Gameloft",
+    "Games2win",
+    "Gearbox Software",
+    "Genki",
+    "Giants Software",
+    "Gogii Games",
+    "Good-Feel",
+    "Grasshopper Manufacture",
+    "Gravity",
+    "Griptonite Games",
+    "Guerrilla Games",
+    "GungHo Online Entertainment",
+    "Gust Corporation",
+    "HAL Laboratory",
+    "HB Studios",
+    "Haemimont Games",
+    "Hanaho",
+    "Harmonix Music Systems",
+    "Hasbro Interactive",
+    "HeroCraft",
+    "High Moon Studios",
+    "High Voltage Software",
+    "Hoplon Infotainment",
+    "Hothead Games",
+    "Housemarque",
+    "Hudson Soft",
+    "Human Head Studios",
+    "Humongous Entertainment",
+    "Hyperion Entertainment",
+    "IO Interactive",
+    "Idea Factory",
+    "Ignition Entertainment",
+    "IguanaBee",
+    "Imageepoch",
+    "Incredible Technologies",
+    "Infinity Ward",
+    "Infogrames",
+    "Insomniac Games",
+    "Intelligent Systems",
+    "Interplay Entertainment",
+    "Introversion Software",
+    "Irem",
+    "Irrational Games",
+    "JV Games",
+    "Jadestone Group",
+    "Jagex",
+    "Jaleco (New)",
+    "Javaground",
+    "Juice Games",
+    "Jupiter",
+    "Klei Entertainment",
+    "Koei",
+    "Konami",
+    "Krome Studios",
+    "Krome Studios Melbourne",
+    "Kuju Entertainment",
+    "Kuma Reality Games",
+    "Kush Games",
+    "Larian Studios",
+    "Legacy Interactive",
+    "Legendo Entertainment",
+    "Level-5",
+    "Lionhead Studios",
+    "Llamasoft",
+    "Looking Glass Studios",
+    "LucasArts",
+    "Ludia",
+    "Luma Arcade",
+    "Luxoflux",
+    "M-Inverse",
+    "Majesco Entertainment",
+    "Marvelous Entertainment",
+    "Massive Entertainment",
+    "Masthead Studios",
+    "Mattel",
+    "Maxis Software",
+    "Mean Hamster Software",
+    "Media Molecule",
+    "Media.Vision",
+    "Mercury Steam",
+    "Microsoft Game Studios",
+    "Milestone",
+    "Milestone S.r.l.",
+    "Mistwalker",
+    "Mitchell Corporation",
+    "Mojang AB",
+    "Monolith Productions",
+    "Monolith Soft",
+    "Monumental Games",
+    "Mythic Entertainment",
+    "NCsoft",
+    "NGD Studios",
+    "NHN",
+    "Namco Bandai",
+    "Natsume",
+    "Naughty Dog",
+    "Ndoors",
+    "Neowiz",
+    "Nerve Software",
+    "NetDevil",
+    "Neverland",
+    "Neversoft",
+    "Nexon",
+    "Next Level Games",
+    "Nihon Bussan",
+    "Nihon Falcom",
+    "Ninjabee",
+    "Nintendo",
+    "Nippon Ichi Software",
+    "Novalogic",
+    "Novarama",
+    "Obsidian Entertainment",
+    "Oddworld Inhabitants",
+    "Oxygen Studios",
+    "Page 44 Studios",
+    "Pandemic Studios",
+    "Paon",
+    "Papaya Studio",
+    "Paradox Interactive",
+    "Pax Softnica",
+    "Pendulo Studios",
+    "Penguin Software",
+    "People Can Fly",
+    "Phantagram",
+    "Pi Studios",
+    "Piranha Bytes",
+    "Pivotal Games",
+    "Platinum Games",
+    "PlayFirst",
+    "Playdead",
+    "Playdom",
+    "Playfish",
+    "Playlogic Entertainment",
+    "Polyphony Digital",
+    "PopCap Games",
+    "Punch Entertainment",
+    "Pyro Studios",
+    "Q Entertainment",
+    "Q-Games",
+    "Quantic Dream",
+    "Radical Entertainment",
+    "Rainbow Studios",
+    "Rare Limited",
+    "Raven Software",
+    "Reality Pump Studios",
+    "Realtime Associates",
+    "Realtime Worlds",
+    "Red Storm Entertainment",
+    "RedLynx",
+    "Redtribe",
+    "Reflexive Entertainment",
+    "Relic Entertainment",
+    "Remedy Entertainment",
+    "Retro Studios",
+    "Revolution Software",
+    "Rising Star Games",
+    "Rockstar Games",
+    "Rockstar North",
+    "Rocksteady Studios",
+    "Ruffian Games",
+    "Runic Games",
+    "Running with Scissors",
+    "SCE Studio Liverpool",
+    "SNK Playmore",
+    "Sarbakan",
+    "Sega",
+    "Sidhe",
+    "Silicon Knights",
+    "Silicon Sisters",
+    "Silicon Studio",
+    "Snowblind Studios",
+    "Sobee Studios",
+    "Sonic Team",
+    "Sony Computer Entertainment",
+    "Sora Ltd.",
+    "Spectrum HoloByte",
+    "Spellborn International",
+    "Splash Damage",
+    "Square Enix",
+    "Star Vault",
+    "Starbreeze Studios",
+    "Stardock",
+    "Sting Entertainment",
+    "Strawdog Studios",
+    "Straylight Studios",
+    "Streamline Studios",
+    "Sucker Punch Productions",
+    "Sumo Digital",
+    "Sunflowers Interactive Entertainment Software",
+    "SuperVillain Studios",
+    "Swingin' Ape Studios",
+    "THQ",
+    "Tag Games",
+    "Taito Corporation",
+    "Take-Two Interactive",
+    "Tale of Tales",
+    "TaleWorlds",
+    "Tamsoft",
+    "Tantalus Media",
+    "Tantrumedia",
+    "Team17",
+    "Techland",
+    "Tecmo Koei",
+    "Telltale Games",
+    "Terminal Reality",
+    "The Bitmap Brothers",
+    "Three Rings Design",
+    "TimeGate Studios",
+    "Torpex Games",
+    "Torus Games",
+    "Tose",
+    "Transmission Games",
+    "Trapdoor",
+    "Traveller's Tales",
+    "Treyarch",
+    "Tri-Ace",
+    "Tripwire Interactive",
+    "Triumph Studios",
+    "Turn 10 Studios",
+    "Two Tribes",
+    "Tygron",
+    "Ubisoft",
+    "Ultimate Play The Game",
+    "United Front Games",
+    "Universomo",
+    "Valve Corporation",
+    "Vanillaware",
+    "Venan Entertainment",
+    "Vertigo Games",
+    "Vicarious Visions",
+    "Virtual Heroes",
+    "Virtual Playground",
+    "Visceral Games",
+    "Vivendi Games",
+    "Viwawa",
+    "Volition",
+    "W!Games",
+    "WB Games",
+    "Wahoo Studios",
+    "Wanako Games",
+    "Wangame Studios",
+    "Webfoot Technologies",
+    "Wideload Games",
+    "Wildfire Studios",
+    "Wolfire Games",
+    "World Forge",
+    "Xseed Games",
+    "YoYo Games",
+    "ZapSpot",
+    "ZeniMax Online Studios",
+    "Zipper Interactive",
+    "Zylom",
+    "Zynga",
+    "id Software",
+    "indieszero",
+    "n-Space",
+});
+
 }
 }
