+++ src/modules/word/WordData.h
@@ -1,19 +1,13 @@
 #pragma once
 
+#include "WordStore.h"
+
 #include <array>
 #include <functional>
 #include <string_view>
 
 namespace faker::word
 {
-extern const std::array<std::string_view, 1328> adjectives;
-extern const std::array<std::string_view, 325> adverbs;
-extern const std::array<std::string_view, 51> conjunctions;
-extern const std::array<std::string_view, 46> interjections;
-extern const std::array<std::string_view, 6659> nouns;
-extern const std::array<std::string_view, 109> prepositions;
-extern const std::array<std::string_view, 5910> verbs;
-
 // https://tristanbrindle.com/posts/a-more-useful-compile-time-quicksort
 namespace cstd
 {
